[roll] Roll fuchsia [pci] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
maintain a tree of devices.  Also removes some uses of "mutable" on
locks that was not needed, and renamed the "device_list_" member of
bus to just "devices_", as this was not actually a list of devices but
a tree.

This refactor removes a multiple uses of custom container traits from
the system which helps to prepare for an up-and-coming refactor of the
fbl:: container template arguments which will require special handling
of users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386825
Original-Revision: 979a5fd5c91d98bd61d050e92e85bebeda3e5d46
GitOrigin-RevId: 5ca0732dfcddae9a6f2e2906e5840ed050ddd708
Change-Id: I73e98344c00f47c1570d5afe2d4cbb5502c31e1f