[roll] Roll fuchsia [build][breakpad] Delete breakpad archive creation
This is no longer needed(!) and neither is the underlying tool that
creates it (dump_breakpad_symbols). In this change, we wave goodbye to
ugly code and hackery (the GN integration), and thank it for its
service. A follow-up will delete ids.txt altogether.

Original-Bug: 49782
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386746
Original-Revision: 9dd7c853048291529d90f9ff187fa469d79e71ed
GitOrigin-RevId: 6e77fc9337f4faeb91c41f5f799db2e6a5a99d5d
Change-Id: Iba0c8b33237d0f13bbf98116c4512bc319169942