[roll] Roll experiences [ermine] Remove login_shell.
Since Ermine migrated to SfW this is now unused and modular is in the
process of removing the base shell interface that it relied upon. Some
day in the not too distant future it will rise again, like a phoenix
from the ashes, in an even more glorious session framework form.

Original-Bug: 48782
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/386716
Original-Revision: 428e5918553ce11b688953ccc2379d5f1b672ebf
GitOrigin-RevId: dc403ff9913256edb0c08e3e68204cfb9f1ba18f
Change-Id: I98a3c0caeef612e9cb1ae766873804e9026fb4b2