[roll] Roll fuchsia [dispatcher-pool] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
implement lists used to manage EventSources in their ExecutionDomains.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386814
Original-Revision: 0267c3eb9e00b9d17a2af8c4c7bab504b76802b7
GitOrigin-RevId: 6b841c856787d2adcee3cc8c8b19227337d7a6c5
Change-Id: Ibc66e3d766acccb9957d598d9880ade9dbc55e32