[roll] Roll fuchsia [fx qemu] add magic switch to qemu to quiet the terminal clear on start
SeaBIOS by default has an annoying output of a terminal clear sequence
on start, causing you to lose your terminal backlog when running it.

Pass an extra switch to qemu to tell SeaBIOS to not use the serial port
at all. The way the kernel is loaded isn't really using the BIOS at all
anyway, so it's not a great loss.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386939
Original-Revision: 021feba55494487863fd23c4c057c70c8bb4886b
GitOrigin-RevId: 23a16ed49b13514d3dbbc59e4c15d828c7c35a46
Change-Id: I918f4bbd4cff4689b4e4db298d0b3735418de393