[roll] Roll fuchsia [blobfs] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
track cached blobfs nodes.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386816
Original-Revision: f8d35c5b15cb2c53b45abc08961f69aa03893ffd
GitOrigin-RevId: e03f26dfb49334b042e0942412ded23f28fe8d39
Change-Id: Ia87a2999d6e8c0827dc253ee0ff82303cdb25249