[roll] Roll external/github.com/catapult-project/catapult Fix Android symbolization race condition
Fixes a race condition in Android minidump symbolization when
symbolizing Monochrome minidumps. Previously, we were simply returning
all possible symbol binary locations and assuming that the correct
dumped symbols would be used. However, in cases of mixed bitness such
as with Monochrome, it's possible to have valid symbols in multiple
locations. This, combined with the fact that non-official builds
typically have null debug IDs for libchrome.so/libmonochrome.so, meant
that the wrong dumped symbols could be used depending on which of the
valid locations was dumped last.

The fix is to check for a directory's viability for symbols for a
particular minidump, and only return symbol binaries in a single
directory.

Original-Bug: chromium:1056235
Original-Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2188954
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Original-Revision: bc56f81214a6f4065cce3d7b656ffddec964ae0c
GitOrigin-RevId: 003d820965cec5d590c42395a46cd8f8506dd4e9
Change-Id: I540042db4a47645154139ea1cf86a8c0ca07e678