[roll] Roll fuchsia [build] Derive unique test name in test spec
Previously the name field was an arbitrary string passed by the caller.
Because it was not guaranteed to be unique, it was mostly ignored by the
infrastructure, and instead rewritten using the logic in this CL:
if component URL is present, use that.
else, use path.

With this change, we can remove this logic from the infrastructure and
just use the name field as it appears.

This also simplifies all the call-sites of the test_spec GN template.

Original-Bug: 50400
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387463
Original-Revision: b3855fd85dd2a63c1fcdeb47a83c5737292eb536
GitOrigin-RevId: dfe33633d2da1bd304c8e8933e413b3485d0fb69
Change-Id: Iee224a2e0287b38a1e32dc6c82b9592bb8975fff