[roll] Roll fuchsia [storage] Update Minfs C++ style.
Add a DirentBuffer class to simplify the task of reading from the device
into a buffer large enough for the largest Dirent and casting it. This
also resolves alignment issues with doing this (although there are a lot
of other places doing unaligned access on filesystem structures).

Use the new DirentBuffer in InitializeDirectory and append the results
with memcpy as a second step. This code is not performance-critical and
the new flow ensures aligned access with less pointer computations and
no casting.

Provide default values for DirArgs. Although not a problem, this class
is not performance critical and default values make it easier to use
correctly.

Rename MINFS_DIRENT_SIZE to kMinfsDirentSize.

There should be no behavior change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388013
Original-Revision: c2482bd2e252b1f2f5388b9e752ae24b05592381
GitOrigin-RevId: 12c63a97f38d87bcc76407abbb91d0b65ed4dbb3
Change-Id: Ie8a617c05b0f27e7a72cf7b9892d70859b1417e0