[roll] Roll fuchsia [benchmarks] Remove non-SL4F-based invocations of most performance tests
Catapult now alerts on the SL4F versions the performance tests and not
the versions being removed here.

I'm not quite ready to remove the peridot-x64-perf-dawson_canyon
builder for the reason explained in the comment, but removing most of
the tests from it will save resources and remove the need for people
to maintain the invocations in src/tests/benchmarks/benchmarks.cc.

Original-Bug: 23091
Test: on CQ: peridot-x64-perf-dawson_canyon builder
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387526
Original-Revision: 73d21dcb0e2a77684fbb4f63696e66acb07fd5b8
GitOrigin-RevId: a6b8cf207bd49ba2c9d48a05c1345ad7f97cb592
Change-Id: I92a8a306667a3d1bba87f67a3a1eeb16176dcd18