[roll] Roll fuchsia [storage] Verify that the tail of blobs are zeroed.
An attacker that is able to manipulate the disk, could change data
beyond the end of the blob. This change adds verification so that this
should not be possible to go undetected.

In theory, this could bring existing corruption that has been
undetected to light.

Original-Bug: 51377
Test: blobfs-unit-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387013
Original-Revision: ec38c12c5caeac7c2714ecbaa6db6c1bf7a734c1
GitOrigin-RevId: 813b7c55afd43a561526c6b47c65711373bd07b1
Change-Id: Icb066a64a7c5b73ed0b5a6bc852a252b4510767b