[roll] Roll fuchsia [bt][sdk][sys] API changes to BondingData
* BondingData now contains a single peer address instead of separate
addresses for each transport. This removes unnecessary complexity when
generating and restoring bonding data for dual-mode devices for which
the transport-specific address is always the same.

The transport-specific address fields are now deprecated. They will
continue to be supported until they can be removed.

* LE data now stores a pair of LTKs (both local and peer distributed)
instead of a single value, to fully conform to SMP legacy pairing
initiator flows. LTK distribution by the initiator is optional in the
specification but required by the ICS when "SMP as Master" is supported.

The old LTK field is now deprecated. It will continue to be supported
until it can be removed.

Original-Bug: 49330, 2411
Test: Compile FIDL
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377228
Original-Revision: b4cb194d7350be2dd9fffb28c404d1df8fc47ebb
GitOrigin-RevId: 8cb1cbd5a24282978e83db9a9138c290d0f2b897
Change-Id: I130c7c114a2524703d9a6819a2f4b894645b98b9