[roll] Roll fuchsia [fidlcat] add --extra-name option
This option allows fidlcat to start displaying events only when one main
process is started and allows to shutdown fidlcat when there is no more
main process.

This is useful when you want to monitor some services on top of your
application. It avoids fidlcat to display the services' events which are
not related to your application.

It allows fidlcat to shutdown when your application terminates (because
services never terminate, monitoring a service with --remote-name
instead of --extra-name forbids fidlcat to exit).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385375
Original-Revision: 7d245d561607c88a679bd77d1889f2327f8cd219
GitOrigin-RevId: 454e295b09aee1e9a64624a571088e3b6122d14b
Change-Id: Ie88916b52ac359c63cdff0e29715167f50d06946