[roll] Roll fuchsia [fidl][benchmarks] Benchmarks for non-byte array/vector
This is to detect when we are overfitting optimizations.
For instance, in Go today uint8 vectors are special cased.

The array/vector elems are non-primitive inline values with padding.
This highlights optimizations that only work when there is padding or
specific primitive types.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388135
Original-Revision: f4d2d24e3d06146b4955e97013e56f13a7ace46f
GitOrigin-RevId: 5b71b5ccc7b14be8dbe6871263c2c1451f84edcd
Change-Id: I15e83bc669d1148d0c0f4575ed2288fa70b8bd67