[roll] Roll fuchsia [memfs] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
track dnode children of dnodes.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386817
Original-Revision: 46d2c84a01275c04fdef1176893428994925b404
GitOrigin-RevId: cc92bfda9679c6f6a11c79f8b3c4a57e290f9496
Change-Id: If5d58f057340112d76b0eda7574bb0cd672792ed