[roll] Roll fuchsia [terminal] Port to render from canvas
This is a somewhat brute-force port of Terminal to render from canvas.
The sole purpose of the CL is to eliminate the last use of canvas so it
can be removed from Carnelian and get rid of a lot of technical debt.

Since the foreground and background color can change per cell, our
usual approach of caching the glyph rasters will not work. Instead
glyphs are created as needed. Once Terminal is unpaused, creating a
sophisticated enough glyph cache would be an excellent task.

This implementation also uses pre-clear, rather than preserving the
rasters to erase them on the next frame. This would be another
excellent task for a resurrected Terminal effort.

It is also possible that these glyph caching and optimal redraw
strategies will end up becoming features of Carnelian and Terminal will
get them for free.

Testing; ran all unit tests and manually tested on core and workstation
builds.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387388
Original-Revision: 02ccbef1b998c284b0ba7e8aa148610493082e26
GitOrigin-RevId: ce90ffa2ccc49b73d781399e0113d9d3020ab93b
Change-Id: I00b07cf39f71a56a1bd672eaba6d4214ccf414ab