[roll] Roll fuchsia [bt][async-helpers] Replace Receiver::receive with Stream impl
Stream + FusedStream interact better with select! and other stream-like
usage patters that will be common with responding channels. There is
also a relaxation in an API constraint where in the previous
implementation only a single request could be outstanding at a time.
This relaxation isn't the motivation for the change, but is
a consequence of using Stream.

TEST: fx run-test async-helpers-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/366358
Original-Revision: 81d74f8fd1bcf8d3be7ce465412d1ef2bc922e81
GitOrigin-RevId: 5460eb95c801306fa8b933123fa39bc1a0c0938e
Change-Id: I0e315baaf9ee98c3a6dc2130ed4edfe250814965