[roll] Roll fuchsia [memfs] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
allow StreamOp and Stream objects to exist in multiple containers
simultaneously.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386818
Original-Revision: cc57ccf4ba00f1325c4deab5536097a72877fbb2
GitOrigin-RevId: 20e1d5729b8375627c6dfab4fad44f2f51a316cb
Change-Id: I4c07a0310591acafb6e7a73259d3f2688e3e4e87