[roll] Roll fuchsia [fidl][benchmarks] Runner for fidl microbenchmarks
This is intended to be used with the chromeperf runners, but is also
useful for running locally and generating data. There is a permission
issue when running the benchmarks (e.g. go_fidl_microbenchmarks) using
this runner on CQ which I am investigating but I wanted to commit this
as-is because it is useful for generating csv data in the meantime by
running locally.

- This currently only supports builder/encode/decode benchmarks
- Produces a standardized output, suitable for tracking in chromeperf
- Can produce csv

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386654
Original-Revision: 2561b986176d6ab6cad7ca342e832e7c8d303785
GitOrigin-RevId: cc3805575bece396ea0a14dbd82f4cde97474c98
Change-Id: I5374781277e8ab54568cd06412b21ba9f8a08358