[roll] Roll fuchsia [archivist] Report "ERROR:" klogs as Severity::Error"
In order to get a better read as to the number of error logs
produced by the system, we need to count klogs as well. Prior to
this CL, all klogs were reported as Severity::Info. This CL is a short
term workaround, until we have structured klogs, to look for the string
"ERROR:" and "WARNING:" to map to Severity::Error, and Severity::Warn
respectively.

This CL addresses a previous issue in the reverted
2ef64fbeae1080e2e15f9698c118d8dce61cea21 patch. Instead of slicing
a string at a fixed byte index, it splits it at a fixed character
index which ensures we handle utf-8 correctly.

Original-Bug: 51179

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387933
Original-Revision: ee5c060e5982d5552a6f76f03bda34ab06103c20
GitOrigin-RevId: ba799c9d78a1e2a44e29d42c4914c447935f4b9a
Change-Id: I9ea322104839040c7f7c58ec14e06042025439ec