[roll] Roll fuchsia [virtualization] Fix virtio-queue index handling
Check that the difference between the ring index and avail index does
not exceed the size of the queue.

In addition, do some more small cleanups in the queue handling logic in
order to be able to remove some of the legacy in-process code in the
future.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388100
Original-Revision: fe52a1fbad412ee0c3dc9c2ea2d87e0d825df744
GitOrigin-RevId: fdc54ecbf93e61047ab87707cc5aa404364ccca9
Change-Id: Icff7b5fad98f590d70248798032cf4b662d92f2f