[roll] Roll fuchsia [perfcompare] Remove support for single-JSON-file "before/after dataset" format
Remove support for outputting this format via the
"make_combined_perf_dataset_file" subcommand, because that is no
longer invoked by the recipes code.  (That used to be invoked by the
fuchsia_perfcompare.py infra recipe, but that has been removed and
replaced by the perfcompare mode of the fuchsia/fuchsia.py recipe.)

Similarly, remove support for reading that format.

This is approximately a revert of
7dbb241b6ee2f5d6d9e1ec1a3cfbdbfb3f3b0a47.

Outputting a huge JSON file into the infra logs turned out not to work
very well.  It was intended to be easy to download with a single
'wget'.  However, outputting large files into the infra logs is quite
slow.  Worse, the large files seemed to disappear from the logs after
a while -- when attempting to fetch the file, the HTTP request would
just hang.  (But I didn't investigate that problem further.)

Fortunately there is a replacement: The new perfcompare recipe
implementation outputs the results via Isolate (see
"perfcompare_dataset_hash" in the recipes code).  This has the
advantage of being quite efficient.  It has the disadvantage that it
requires a special infra tool to download the data (isolateserver.py
from client-py).

Original-Bug: 49280
Test: "python perfcompare_test.py" (run by bots on CQ) + perfcompare trybot
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388316
Original-Revision: 375ccf0e65925cbb8c74ee12c6b1ea80baecbfd9
GitOrigin-RevId: 82805e146228325d743394d0d1aab90f77eeb652
Change-Id: I38e96f33764fbfa97b97a17bf9ec27e5e8dc58c8