[roll] Roll fuchsia [cobalt][test_app] Utilize the new ListenForInitialized() method.
This fixes a race condition between the Cobalt test_app and the
initialization of the Cobalt service by replacing a sleep
with waiting for the initialization signal.

Original-Bug: 51211
Original-Bug: 51473.

This does not yet fix the Prober. In a follow-up CL I will
fix the Prober by stopping using the FakeTimekeeper with
the prober.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388256
Original-Revision: bc972efb98cdde95678ab818217d0944333af379
GitOrigin-RevId: 6f373e5ccfc45972756cf7ae61688ea992bb756c
Change-Id: If1bc993315aa080fee58d878d59279804bcd8617