[roll] Roll fuchsia [sysmem] Add API for setting BufferCollection name
We want participants that don't receive VMOs to be able to name them,
because they might have better high-level ideas of what the VMOs are
for. So add a SetName method to buffer collection to allow any
participant to set the VMO name. This also allows passing through the
name to the Allocator, so it can print out buffer names when it runs out
of space.

Add support to magma and the amlogic video decoder drivers for setting
this property, so we can take advantage of the new feature immediately.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387389
Original-Revision: ea478e7270911efc06e07e9e833b9a4bb6852790
GitOrigin-RevId: 3ed7e96f36087e89a05e1261668d2c6fe9218d81
Change-Id: Ib44e72c1b1975d9beed80c86499673077e4a289a