[roll] Roll fuchsia [wlan][iwlwifi] port iwl_trans_pcie_log_scd_error()
This function is used later to dump the Tx queue info to debug
in the error cases. Make it standalone so that easier for review.
Most of changes are style.

Original-Bug: 45032
Tested: Compiled only. Will be used in the following CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385380
Original-Revision: abd8f916ffb56b7709e65bb787db5c3ca7c842e6
GitOrigin-RevId: bcbd47025c536e45ddb0ad8a6c0792e2349228c5
Change-Id: I7a08ee0686a4cd2cb41220bd2f0de376a6139efe