[roll] Roll fuchsia [ffx][target] Refactor Target to avoid deadlocks.
Refactors some code around the Target struct in order to make it more
difficult to create potential deadlocks. It is now not possible outside
of testing and inside of the target module for there to be more than one
lock acquired simultaneously in a Target struct.

It was not feasible to create a reliable public API test that would have
(previously) created a deadlock without simply acquiring several locks
in step with each other (which is not possible now with this change),
so this only includes tests for newly added functionality. The remaining
code is a refactor already covered by tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385362
Original-Revision: 1cab74880e92d77b0f9530acd6dd46e23b640682
GitOrigin-RevId: 6ba8d7442fea9a6df0ac574ecc6dc43e98ca3b11
Change-Id: Id3b28c57b7f4a3d104f90709b10bc02fe70ed832