[roll] Roll fuchsia [bt][lib] Use `peer_ltk` and `local_ltk` fields in Rust library LeData
The fuchsia-bluetooth Rust library LeData type now contains the `peer_ltk`
and `local_ltk` fields which replace the old `ltk` field:

- The conversion functions for sys.LeData still support the deprecated
`ltk` type. When the new fields are not present, the conversion defaults
to the deprecated field.
- The control.LeData type only has a single `ltk` type. The conversion
functions currently map this field to both `peer_ltk` and `local_ltk`
when converting from FIDL. The fields are assumed to contain the same
value when converting back to FIDL. This change is compatible with the
current bonding storage format and bt-host, both of which currently use
the control FIDL types.

Original-Bug: 2411
Test: 1. bluetooth-crate-tests, bt-gap-tests, bt-integration-tests
      2. Tests that pairing and re-connecting following a reboot still
      work with the new format.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378680
Original-Revision: 1bc5651471174d2c0451257f8cef1ec44dabd889
GitOrigin-RevId: 1b93c5e4a50d87f817acf9504591a5caec46a4ff
Change-Id: I40d2d19a3997c4156e5f36127ea5281299ceaa0a