[roll] Roll fuchsia [e2e] Creates a no-basemgr test
The "Screen is not black" test is intended to be ran on a device
configuration that supports taking screenshots, so that we can establish
whether there is something shown on screen when the time comes.

Some system configuration had basemgr running at boot, and some
configurations wer able to run basemgr, but didn't have it running.  The
test has code that tries to figure out whether basemgr is running and
tries to start it if it isn't.

This approach does not work on system configurations where basemgr is
not supposed to be running in the first place.  So, in this change, we
add a separate "no-basemgr" test, which assumes that the device has
another way of obtaining its screenshot and simply skips trying to
fiddle with basemgr.

This is not the most general approach, but based on what has been
reported in bug 50705, could be enough to make the test run well
in configurations that are currently of interest.

A side change, now that I am here, added canonical targets for the
directory and tests.

Original-Bug: 50705
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386238
Original-Revision: edfe28abbdf3b84963634364570f148cdc1c685e
GitOrigin-RevId: c86ab3c6d14d61ddcd42735448847b475df14a1c
Change-Id: I3e134ba8decd0defea4d4d79b76ef42c3fefeaf5