[roll] Roll topaz [flutter_test] create a unique package_root
The newest version of dart requires each package to have a
unique package root. The flutter_test target creates a separate
package for each test source which will break in a subsequent
roll of dart. This CL will create a directory for each source
allowing it to act as the package root.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/topaz/+/388802
Original-Revision: 50a3457f624553ce76ca3d1526c819959816152d
GitOrigin-RevId: 5012ef2600aeb203f626bfaec1d65ea11fa1eb9e
Change-Id: I4bb4aff20e07363f43eabf6674f5b15e82571ca3