[roll] Roll fuchsia [intl] Adds a build template for generating l10n
Adds a template `//build/intl/strings.gni:intl_strings` which takes on
most of the responsibility for generating localized resources from
localized material.

Since the template uses FIDL under the hood to generate polyglot
includable code, and given that this seems to be the first such use of
FIDL, I had to tweak the FIDL templates to allow `non_fidl_deps` to be
declared.

Since this all is a bit cumbersome still, I added an example to show how
one can deploy a Fuchsia package which compiles with the generated code
and bundles the localized resources in the Fuchsia package.  The
example will evolve with the sophistication of our L10N code.  The
example could probably use some trimming down, but at the moment I am
not sure about the approach.  So I left that for some followup work.

Original-Fixed: 51172
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386391
Original-Revision: cda2a0365531ae61bba926092cb29da543ffa1f0
GitOrigin-RevId: 6f91c582b42b0d06452cd5298d18589c713db4da
Change-Id: Id71fc87cedfae15da4ba3b2c80f0b1d19373fb07