[roll] Roll fuchsia [bt] Create data domain fuzzer
Create an overall fuzzer of the data, hci, and l2cap stack.

Fuzzer data is received as ACL data packets and also triggers the
opening and closing of a connection (so that some packets are valid).
One service is registered to enable the possibility of the fuzzer
opening an l2cap channel.

Don't fuzz the length field of ACL data header because it is set by the
controller and we are confident we handle it correctly. Most fuzzed
length fields would be incorrect, and so most packets would be rejected
immediately, which would slow down fuzzing.

Don't fuzz the handle field of ACL data header in order to speed up
fuzzing. Use handle of connection that is toggled by fuzzer instead.

Original-Bug: 51794
Test: (only adds test code)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387759
Original-Revision: e66beae5c86ecd221ac079bdf4c5384091a48b0e
GitOrigin-RevId: 95e9a246e3dfb8d35c501f463b99b73a9cc46b8f
Change-Id: I0cf929fce31f78cf22582fa7f6d862a3627189ae