[roll] Roll fuchsia [kernel][x86][kasan] Use sanitized memcpy/memmove/memset (5.5/N)
This commit makes asan builds to use `__asan_memset`, `__asan_memcpy`
and `__asan_memmove` as the default versions for `memset`, `memcpy` and
`memmove` respectively.

The change defines the memcpy/memmove/memset symbols as weak, and
defined a strong alias in `asan-stubs.cc`.

This CL modifies the memset/memcpy patching logic so it references
the unsanitized version of the memset/memcpy functions. It also adds
unsanitized versions of the functions for arm64.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388361
Original-Revision: d727c89acfa3d4004d340713d24c09874606ef69
GitOrigin-RevId: 655f98abefcda76188f5f96378f4f4e73c79e11f
Change-Id: I9dc915496a62fb661e52a56131283605da440df9