[roll] Roll fuchsia [fuchsia.hardware.input] Add DeviceReportReader
Add a new interface that reads reports. The
new interface allows the driver to be in control
of the new interface and have a HangingGet
function. It will also allow the instance drivers
to be removed in a later commit.

Take this time to have a Report struct,
so that a vector of Reports can be returned.

Original-Bug: 49329
Test: hid-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386295
Original-Revision: ad01e35686efc3bf70365c024b8c56408fbc5863
GitOrigin-RevId: 7887975d3eb542ce9fc19ad91d8370c350a5bdc1
Change-Id: Ia8f6a9ef89614339284668d5cb6cd32439854e0a