[roll] Roll fuchsia [hid-input-report] Use LLCPP BufferAllocator
Use the new LLCPP BufferAllocator to allocate
DeviceDescriptors and InputReports.

The hid-input-report driver now stores one
large buffer which it allocates from to
create InputReports. This buffer is
cleared when the reports are sent to
a client.

The hid-input-report driver's device
descriptor is now stack allocated.

The old functions are going to be removed
in a later commit.

TEST: lib-hid-input-report-test
      hid-input-report-test
      Manually tested on hardware.

Original-Bug: 49628
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382273
Original-Revision: 895433b48621bec97191344cbf9ca6dcf8519680
GitOrigin-RevId: c041f5f7c42ca87dad3f29465bd9406c5f64f952
Change-Id: I622e54b776ba31c6c1a984c58d4209f56a17a620