[roll] Roll fuchsia [blobfs][rac] Bug fix: num blocks calc: ZSTDRead()
This change contains two tests and a bug fix. The bug is an incorrect
calculation of how many blocks to read in `ZSTDRead()`. This calculation
is currently based on the size of the read. It should be based on the
block-aligned offsets in the blob of the start and end of the read.

For testability, `ZSTDSeekableFile`, `ZSTDSeek()`, and `ZSTDRead()` are
now public, rather than hidden in `zstd-seekable-blob.cc`.

BUG=50837
BUG=44820
TEST=`fx shell runtests -t blobfs-unit-test`

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383641
Original-Revision: 5d7e59b88257d435cc8f47a03c393c18ec504fcb
GitOrigin-RevId: 152283e68a08d259575bdacb799bdb47db685ead
Change-Id: Idc25768b467ba854c1ca503050094d331f42b8ba