[roll] Roll fuchsia [Harvester] Reduce warning messages to verbose level
The Harvester now caches a list of jobs/processes/threads to amortize the high
cost of gathering the list. This means that items in the list can be out of date
(increasing as the cache ages) and lead to more spamy log messages about
failing to get task information. This CL changes the log level of those messages
from WARNING to VERBOSE3. So the messages can still be seen if verbose3 logging
is requested.

bug=48250

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386253
Original-Revision: 97f07a7d0a1cdaea724fd679334cf73f43994092
GitOrigin-RevId: 634118c0ccb47ccf3681e8a1ffbaf9ada180ef75
Change-Id: I882506d770e96d51ddb779393907f629982c8756