[roll] Roll fuchsia [ddk][devmgr] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in various
places in the devmgr.

This refactor removes a multiple uses of custom container traits from
the system which helps to prepare for an up-and-coming refactor of the
fbl:: container template arguments which will require special handling
of users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386824
Original-Revision: a016047a2a4f5b1742fbbecc3aaa1b8a1b8b1611
GitOrigin-RevId: db557b5ca8028365b78af8dd50150f692554fc29
Change-Id: Id5986580fe0c1fe8e9842ed418c2644a6ccc43de