[roll] Roll fuchsia [dhcpv6] Define FIDL API for the client
Acquiring DNS addresses in stateless DHCP is the first use case we want
to deliver, and this API enables that.

`ClientProvider` is kept minimal. This way if in the future we want to
run one client per component, the change needed to `Client` protocol is
also minimal.

In the future we'll extend this API in 2 ways:
1. Allow starting DHCPv6 client in other models
2. Compose other watchers, e.g. an AddressWatcher

Original-Fixed: 48861

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387737
Original-Revision: 2ecb0be58b76fadca1de755bad953001219025bb
GitOrigin-RevId: ff6c09e43066796cee60d4deb24164e7d73ef812
Change-Id: I458a168fd2caa4a2860d19066bd34a522eeb21f1