[roll] Roll fuchsia [soundplayer] fix race discovered in test
When soundplayer plays a sound, it employs an AudioRenderer. Among
the methods it calls on the AudioRenderer are SendPacket and Play.
The response to SendPacket that indicates that the sound has finished
playing.

Soundplayer is supposed to wait for both SendPacket and Play to
respond before responding to PlaySound. The code that does is has
a looped select! macro rather than just a join!, because the client
may ask playback of the sound to be stopped asynchronously. The
SendPacket and Play methods are joined using a mutable boolean that
indicates one or the other methods has returned. If one of the
methods returns and that boolean is set, it's time to leave the
loop.

The boolean was set to true initially, which means the loop broke
when *either* method returned (almost certainly Play). This means
that PlaySound was returning early, a fact that was detected by
a test, albeit intermittently due to a race.

This CL sets the boolean to false initially, so the manual join works
as intended.

Original-Fixed: 51871

Test: fx shell soundplayer_bin_test
      The existing test found the problem.

Multiply: soundplayer_bin_test: 30

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389334
Original-Revision: 3a474c8a8200b8e330fbb978f2a45d3a333035e7
GitOrigin-RevId: 83e1e38bc481aa2db4474bf64d85009747acd059
Change-Id: Iaaaf33e8fbe7a0dd6d4becfdeccdfc91dacab22b