[roll] Roll fuchsia [intl] Examine the assets directory
Adds the code that examines the assets directory and returns
the locales found there.  This is one of a few things needed
to provide asset loading on demand.

This was made to be a self-contained change for ease of
review.

Original-Bug: 50725
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387476
Original-Revision: 5707b5781887933b0b44661c2f45be611d54c888
GitOrigin-RevId: 17469f576eb4feced71baacb97f63eaf8ba88fd7
Change-Id: I26cc19028444880b0ee9196e20dbb68f3c0cb3a7