[roll] Roll fuchsia [intl] Factors out Model from json
Should be a pure refactoring.

json crate used to be intertwined with parser, which in turn is using
host-only libraries.  And rightfully so, since we only ever want to
prepare host tools on the host.

However, the loading of JSON libraries and their interpretation is also
the task of the code on the target.

This commit factors out the respective code and makes some modifications
that make this change possible.  `Model::from_dictionaries` had to
become a stand-alone function in `json`, because it needs host-only
`parser::Dictionary`.

Original-Bug: 50725
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387525
Original-Revision: 5bfd53b53d46f39b49e49feffebf17f8a7f7d701
GitOrigin-RevId: f29321396a63d3dc75ba570863b083b05c78c4d3
Change-Id: I8f5dc50fc2bba30eb00cb1195e4a80a3881cb7ff