[roll] Roll fuchsia [fs] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
allow PseudoDir::Entry objects to exist in multiple WAVL trees
simultaneously.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386820
Original-Revision: 4a4c04c56b1269283dfc0ba26cda02c216de8456
GitOrigin-RevId: 14b9b7ab517cf684dbb33ac7f1e659544d192a8f
Change-Id: I68b1e79b9e4ec641d6136cb932d64633d2d51f83