[roll] Roll fuchsia [kernel][vdso] Sanity-check the vDSO image matches the kernel
Extract the build ID note contents at build time and embed that
separately in the kernel so it can verify at runtime that it
matches the vDSO image.  The kernel bakes in expectations about
the exact instruction sequences that appear in the vDSO, so it's
crucial that it always be the precise vDSO image that the kernel
was built for.  The image is pasted into the kernel binary at
build time today, but in the future it will be packed differently
where accidental mismatches will become possible if the wrong ZBI
repacking shenanigans are done.

Test: change consists of adding an assert on the boot path
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383644
Original-Revision: f4bb364fa4963a45cb977a3910dfe05f80cf6bd0
GitOrigin-RevId: cc20b03a44f55693e4afe3ef0d35b7e356983914
Change-Id: I7b60d18aa0b30b0bf308ed9e46f78f7611a85504