[roll] Roll fuchsia [debugger] Skip compiler generated code.
If a backtrace or stop position is on an instruction marked in the line
table as compiler-generated, previously we would show "no line info" in
the stacks and assembly in the stop information.

This changes the Address->Line querying function to return the next line
in the line table when this happens.

This does not change the stepping logic because the line-table lookup
for that is a completely separate code path that calls into LLVM
directly. That code already handled these "line 0" cases properly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389337
Original-Revision: d1c3d80f366e0e8ff1bc15c874874d930a26ba99
GitOrigin-RevId: b3fa6a3680b4ef901407cb5b96ea141ed6f77a18
Change-Id: Ie6f77ba52119ec819de2ef5fb50911c592fc85ee