[roll] Roll fuchsia [fidl][rust] Report epitaph on write failure
Due to the shape of the API, epitaphs can only be reported
when a FIDL method that expects a response is sent.

Fire-and-forget FIDL methods are not `async` so there is no
way to wait for receiving an epitaph. An API change is needed
to support this.

Original-Bug: 50730
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386377
Original-Revision: 089fa4b1dd8cae5c521d150bfdbc4b92c251be0c
GitOrigin-RevId: 26de3f9f1c249b174385ae33864bf81465f47553
Change-Id: I4a11a602035d1ebacc3d9fe32808d8c207bbfbec