[roll] Roll fuchsia [update-package] support packages.json version as int or string
This change modifies the update-package crate such that packages.json
files can have a version that's either a string or an int. I kept in
some of the "old" code so that it's easier to revert this in fxbug.dev/50754.

Original-Fixed: 50713
Test: fx test -o update-package-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389348
Original-Revision: 23e35367e71f93295c8c141012cbb500f5672fa5
GitOrigin-RevId: 144e2317165191c126864635ecddb8794f3bb4cc
Change-Id: I80e938cfec65dfee92cb9712697880d07c0c4205