[roll] Roll fuchsia [blobfs] Support paging chunk-compressed blobs
Implements basic demand paging for blobs compressed using the chunk
compression library (//src/lib/chunked-compression).

Load the seek table for a blob when it is first accessed.

On a page fault, the seek table is consulted to find the compressed
frame containing the target page. The frame is decompressed and all of
the bytes within the frame are verified and served to the client.

Tested: Manual and unit.
Tested: Added variant of integration test.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381177
Original-Revision: 5facb5f7b2c7d69c3e2d7adb86d611b6a804c222
GitOrigin-RevId: d642062452f020ff25dc36662c5b2c3b92009ea0
Change-Id: Id275e93299cba696c62c331c8058b175bc3a7780