[roll] Roll fuchsia [sysmem] Add index to vmo names
This makes it easier for clients to identify which VMO is leaking, if
only a subset of their VMOs have problems. Also add the koid on to the
logging so that we can correlate sysmem's information with that from
memgraph -v.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389217
Original-Revision: f4aef2de07e7a741a50f3fb409e6a0f4f1594d05
GitOrigin-RevId: 6252a8f654b79e153f24e5853bdae647c26ab183
Change-Id: Id3d0716ea1a488c24e18c227e25d7d882f9156a3