[roll] Roll fuchsia [modular] Improve logging around terminating session agents.
1) Log when we signal an agent to perform graceful termination.
2) Log when an agent has terminated expectedly during shutdonw.
3) Change the language around unexpected terminations & remove the
zx_status_t, as it was the status of the AgentController channel, which
is generated by Modular, not by the agent, and is misleading.

For (3), we should revisit this in the future to capture the
OnTerminated status from the ComponentController.

Test: existing
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388119
Original-Revision: ccb5f9bb3ddd2187c7849aaa183204f7dbb001d2
GitOrigin-RevId: 1bef72bd623e2f02789a528a7de1f1ed6723df45
Change-Id: Iee529233e08f388c28d54d9d2a7254467fecfd43