[roll] Roll fuchsia [rust_crate] Move rouille from target to target + host.
Rouille provides a simple web framework that is also applicable to host
tools. This framework is planned to be used in an upcoming host tool so
moving it so it can be used both on the target and host.

Original-Bug: 51992
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389393
Original-Revision: a4dfd3a1b2f626dbd0b7787583f5c5c8dc349fae
GitOrigin-RevId: 15d6fcbf684e52ade716bafa2efa6d7591c3ea8e
Change-Id: Iee67b4fffe0b43225bd79f9370e83797312a5c95