[roll] Roll cobalt [Cobalt 1.1] Add registry fields to support differential privacy.
Add the new PrivacyLevel enum which maps to epsilon values instead of
noise parameters.

Add fields for the RAPPOR noise parameter and discretization parameter
to be computed when the registry is parsed.

Add bounds on values to be reported with privacy.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/387999
Original-Revision: 57d183ea5638e3e2dc4b2540bc08e3a604350e5b
GitOrigin-RevId: 522cbcf175dc8089bf3c7ba9cc109b3c82696914
Change-Id: I2320f8e93e169ed4f0f09abe7a7baaee6a5e94d5