[roll] Roll fuchsia [virtio][socket] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
allow SocketDevice::Connection objects to exist in multiple containers
simultaneously.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386822
Original-Revision: 1c3e2bdb52627735be47e9aa26c32a7e4e120005
GitOrigin-RevId: 0fae3bc2235b9a001d1fb674b4626aab3b407397
Change-Id: I1070070e582fe3cf2b31944168831bae7da2090f