[roll] Roll fuchsia [fidlgen] Delegate fidlgen_syzkaller ownership to Fuchsia security
This CL grants the syzkaller-on-fuchsia maintainers ownership of
fidlgen_syzkaller, so we can freely iterate on this code since we're
the only users. Leaving it in place for simplicity, and so fidl team
can continue making any fidl-wide refactorings as appropriate.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389934
Original-Revision: 6f67b855a041b4d312da00987856f509fa04ac9b
GitOrigin-RevId: a2bd5030d7355a421c8e7ae51489043ddf2e4628
Change-Id: Ia36f7d28598a988115446d4d99281403a7dc7d39