[roll] Roll fuchsia [fdio] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
allow LocalVNode::Entry objects to exist in multiple WAVL trees
simultaneously.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386819
Original-Revision: 22845d6a3a57ff6ca3cb19bf722646ab5769a2d7
GitOrigin-RevId: b996ce0a88bc551c05d393cca20aa3eae13b5724
Change-Id: I0afc26eb8945b5d345b04e7f3656c402f770a4ac