[roll] Roll fuchsia [carnelian] Demonstrate rotating app content
Modified the button sample to draw with one of
four rotations as an example of how the recovery
UI might accomplish a similar task.

Used this rotation requirement as a chance to
experiment with Euclid's typed geometry. So far
it seems to be a worthwhile feature, since it
helps prevent the mixing of display aligned and
presentation aligned coordinates.

While this change will allow recovery to present
its UI correctly, more design and implementation
will be needed to make apps in general function
correctly when running on the frame buffer on a
device with a rotated screen.

Testing: tested the button sample on the frame
buffer as well as under scenic. Ran all unit
tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383480
Original-Revision: 1695d028bfd9a311bcc4875763a74b18b68b7dd6
GitOrigin-RevId: fda78c26291685b12da4516f19a496fb95bc699e
Change-Id: I75c72b24acdc62b0f69f8898dc86dc34dd63bc02