[roll] Roll fuchsia [netstack] Log DHCPv6 observations to Cobalt
...logging the events observed since the last time we logged to
Cobalt, while keeping polling logic outside the Cobalt client.

main gives the ndpDispatcher a `hasEvents` function to call, which
registers novelty with the Cobalt client.

When the Cobalt client is ready to log, it collects events from the
producer, allowing the caller to ensure that all novelty since the last
log is included without coupling the Cobalt client to the specific
business logic for that metric.

This ensures that we only log when we have to, and we only log what we
have to.

Original-Fixed: 48858
Tested: fx test netstack_test -- -test.run ObservationsFromDHCPv6
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/374937
Original-Revision: bd808de38382388ef6f525310f42c7d23e79e7c0
GitOrigin-RevId: a04ef5ac0f0893587fc27e70673eb500f39eabd3
Change-Id: I0f16705103a24749c52c0893bd751b4997403f38