[roll] Roll fuchsia [blobfs] Offload read-only blob connections to a zx::stream
After this CL, read-only connections to a blob are offloaded to a
zx::stream.  The zx::stream operations are still performed in the blobfs
process by libfs.  A future CL will offload these operations to the
client to improve performance.

This CL is a rebase/reland of
https://fuchsia-review.googlesource.com/c/fuchsia/+/369660

Original-Bug: 45691
Test: No behavior change.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375441
Original-Revision: 4a83d164c822fbb12267def25ff5f8c4c144f93c
GitOrigin-RevId: 088d8264c0d9c09854f33f80837bbaf8365209c8
Change-Id: I4d7d29f2d968360024dcb7fd92a170568ee94fff