[roll] Roll fuchsia [intl] (Finally) implement the l10n message lookup
The implementation is no longer a dummy one: it will return
the translated message given the correct message ID.

The implementation is simplistic in that it assumes we know exactly
the list of languages we want to support.  This should be enough
for now.  Locale fallback is currently not done programmatically;
rather, we will package every locale and use pkgfs deduping to
remove ambiguities.

Original-Fixed: 50725
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388118
Original-Revision: f4d25c3c0f912c2d6fb0ffbf02ad1b7bd9345a5f
GitOrigin-RevId: c36fc6cd41000a8995c75bb0ca9af07f35802605
Change-Id: Idb0d4c48f81c8d1cadd4831267ba8798dea6dba6