[roll] Roll fuchsia [bt][bt-gap] Return Ok when access service terminates normally
We erroneously returned an error value when an Access service provided
to a client by bt-gap terminated, even though this is normal behavior.
Instead we return Ok if the task terminates because the stream closed
normally, reserving errors for if an error is received handling a
particular fidl request.

Test: fx test bt-integration-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388594
Original-Revision: 6e6eeaaddcaf79950288b8c04b438d85f70fb67f
GitOrigin-RevId: 8a4af724402cbcc6d3174be41d30d0302193edc3
Change-Id: Ie6e18f5e9307ddcfae69cdb1cc30f12d89045c22