[roll] Roll fuchsia [bt][l2cap] Fix ERTMRxEngine tests, clean up docs
Fix bug in ProcessPduCallsRemoteBusySetCallbackOnReceiverNotReady and
ProcessPduCallsRemoteBusyClearedCallbackOnReceiverReadyAfterReceiverNotReady
that checked for the ordering between RemoteBusy callbacks and the
AckSeqNum callback, which is not relevant. The latter callback is for a
local variable when it should be checking against the ReceiveSeqNum
callback which is for the peer variable.

Additionally, add docs and rename some tests.

Original-Bug: 1353
Test: in bt-host-unittests, update
L2CAP_EnhancedRetransmissionModeRxEngineTest.*

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389963
Original-Revision: 09a5a318a2e5373885c250d16530b54d2b0811ef
GitOrigin-RevId: 5a5ae99cf018c298ae6dc7ba36d2634ed0a3843f
Change-Id: Idb06c2f0cf6a4189657f1bf2b5a5353af56bffda