[roll] Roll fuchsia [debugger] Support breakpoints conditional on hit-count
Conditional breakpoint is a breakpoint that only stops the execution
when some condition is met. It's useful in a loop where a breakpoint
could be hit thousands of times. This patch adds support for "hit-mult",
which means the breakpoint will only take effect when the hit-count is a
multiple of the given number.

An option for "break" command is added, together with a setting called
"hit-mult" which can be get/set on breakpoints.

Original-Bug: 43929
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389614
Original-Revision: beea4fd7d8d57ea798295e440a1c4657a02697d7
GitOrigin-RevId: ae47a3025e90bb86d1ce8bd852c9a19f0488487e
Change-Id: Icda5420dd7851e5d654b3e3869a14a5471cb89a9