[roll] Roll fuchsia [ffx][overnet] Cancels overnet pipe if host-pipe command fails.
This adds additional tracking on the host pipe process to ensure that it
is cleanly torn down (including the reimplementation of std::io::copy
running in the background to helpn it communicate over FIDL channels).

This change came about as a result of debugging hanging issues that
originally seemed to be part of the daemon itself.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386936
Original-Revision: 83302f1b74bf775f834522590fd006d67c489b95
GitOrigin-RevId: 9f07f266380c3025d63950beb0a50d39762388d9
Change-Id: I585d03cdba571305cd52a21bbf39b76e41fd7436