[roll] Roll fuchsia [bt][host] Fix race condition in RunTaskSync()
Fix a race condition in RunTaskSync where it was possible for the
condition variable wait to be spuriously woken up *after* the done
condition was set but *before* the call to .notify(), such that the
condition variable did not live long enough and the call to .notify()
would crash.

I believe this is responsible for flakes in bt-integration-tests as seen
in fxb/49570.

Test: fx test fuchsia-pkg://fuchsia.com/bluetooth-tests#meta/bt-integration-tests.cmx

Original-Bug: 49570
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385269
Original-Revision: 6fcb01c0067d7feebb8c4b12dbc9c0f4a59380fa
GitOrigin-RevId: 2362729219ac52bfbdc24f11ee86ce3b689ab109
Change-Id: I55e3a36f4617044062a56eaeae6f8eb6f2b8440c