[roll] Roll fuchsia [blobfs] Avoid inflating during compression and reduce threshold
This teaches the target and host side compression logic for blobs to
never write a compressed blob if it would be larger than the original.
We store a bit for each blob indicating whether it is compressed so
there's no reason to inflate a blob if the compression algorithm would
otherwise want to. The zstd algorithm will return larger outputs for
some small inputs in the configuration we are currently using.

This also reduces the minimum size for considering compression on a
blob to a single block, as we do benefit from compression for some
small files.

Test:
- Host compression behavior tested in blobfs-host-test
- Unit tests for small compressable and incompressable blobs in blobfs-unit-test

Original-Bug: 39667
Co-authored-by: James Robinson <jamesr@google.com>
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389473
Original-Revision: fc9a72c5237c27b413e5cbe6cfe404a6f1cec9b0
GitOrigin-RevId: c352f55179e7f121c9bb811453ce631d1adaee2b
Change-Id: I316555ac548f5facd537df864e93aaa5cd54e0bb