[roll] Roll cobalt Give a non-null internal_logger to the uploader classes.
Original-Bug: 52303

ClearcutV1ShippingManager and clearcut::Uploader both
need to be given non-null internal_loggers so that
they can participate in Cobalt-measuring-Cobalt.
This is for the metrics:

per_device_bytes_uploaded
per_project_bytes_uploaded

I also needed to make a few changes to allow this to be testable.

Other miscellaneous changes:

- Previously, the constructor for ClearcutV1ShippingManager took
an optional parameter |internal_logger| which defaulted to nullptr.
The problem with this is that it is misleading because if
I were to pass a non-null internal_logger that would not
be used to initialize the internal_logger of the instance of
clearcut::Uploader owned by the ShippingManager.
On the other hand invoking the ResetInternalMetrics() method
 does also update the internal_logger of the clearcut::Uploader.
I decided to make this more clear by removing the parameter
and declaring that you must call ResetInternalMetrics() if
you want internal metrics enabled.

- CobaltConfig::shuffler_encryption_key() and
analyzer_encryption_key both return a std::optional<std::string>.
But previously those methods were implemented by returning
a copy of a std::string member variable. So there was no
point in having the signature be std::optional<>. This
causes a problem if the values are the empty string and
are used to try to create a real EncryptedMessageMaker because
the real EncrptedMessageMaker constructor will fail if given
an empty string. There is code in CobaltService that checks
the status of the optional<> returned from these getters
and uses the result to decide if a real EncryptedMessageMaker
should be created. So we want the return value of these
getters to yield false when the strings are empty.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/390620
Original-Revision: 70807b0f664a972ab8f2aaad3ef24a6dd57df14a
GitOrigin-RevId: b17774a0945da90571a3908b87d32dcd11d14f67
Change-Id: Ia63bde17a89d5d95838b1f09796383eec27f761f