[roll] Roll fuchsia [gn sdk] Update component.gni to track dependencies correctly
This correctly sets the dependencies on the manifest and generated
metadata file used to create the package manifest.

Additional tests were added to cover:
 * overriding the executable output name. This requires the addition of
   the "binary" parameter to the component rule, otherwise it gets the
   name from the executable passed as a data_dep. It is a passed
   as a data_dep so that runtime dependencies are passed along to the
   package rule.
 * touching the component manifest file should trigger a packaging.
 * changing the name of a component manifest should trigger repackaging.
 * Additional targets in "chromium_compat" folder which capture the
    current state of component and packaging rule usage in Chromium.
    This is used primarily to make sure the changes are backwards
    compatible with Chromium. Once the GN SDK rules are adopted by
    Chromium, these targets will be removed.

The result is that changes in the properties or files that are
inputs or affect the building of a component cause the component
target to be rebuilt, which in turn, results in the package being
rebuilt.

Testability: This tests the functionality by adding a commandline arg
to the BUILD.gn file to override the name. This test cannot use
multiply since it runs on the SDK trybot and does not use the test
framework.

Original-Bug: 47991
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382321
Original-Revision: 9294f2aa9e5703f487894783ffb8a96bdb4a894e
GitOrigin-RevId: 1e12d2ae8af1e7d9541992b2e6c074b803514516
Change-Id: Ib96b6ac94ea3a7d065103f9afc9404735d3107f8