[roll] Roll fuchsia [size_checker] Add separate budget for shared libraries we distribute
The Fuchsia project builds and distributes a number of shared libraries
to partners through the SDK.  We also change the contents of these
libraries periodically.  Since they're coming from us, it makes more
sense to budget the size for these separately from the package or
packages that may use them. This allows a size budget specification to
define the set of file paths in packages that should be considered in
this category and an overall size limit for these based on their
expected size and propagation times.  If a dedicated size limit for
this category is not present in the input specification, this check is
ignored.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390640
Original-Revision: 493a186ac0cbf2b94230c9021362d4a39b4cf5bc
GitOrigin-RevId: fb1a82e762e224f45de11d78b5aec0ff817a6256
Change-Id: Ib4d23d1d82953650b718b846f4355c62713dc514