[roll] Roll fuchsia [netstack][test] Disable retransmit timeout test.
This test is waiting for the connection to timeout after multiple
retransmissions. After a recent change in upstream gvisor, this
timeout increases to ~20mins which is unreasonably long test time.

The TCP connection times out when there are N retransmits (15).
The retransmit intervals exponentially increase with a lower bound
of 1sec (or 200msec if there is a data-ack) and upper bound of 120sec.

We can enable this test when we have a way to tune the N value above,
thus reducing the test duration.
Similar to Linux sysctl net.ipv4.tcp_retries2

Original-Bug: 52278

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390534
Original-Revision: d2eda6ada45972ad0e22b5f24d6c2477f2b519a1
GitOrigin-RevId: e803faa204458e54149c08d3de7118f4cb803012
Change-Id: I8b8fb79d1c01198461d431d0a3217eb532e709b6