[roll] Roll fuchsia [fbl][dll] Add AllowRemoveFromContainer option
Introduce a new node options "AllowRemoveFromContainer".  When set on
a NodeState or Containable mix-in, this will allow users to directly
remove a node from a container without needing to have a reference to
the container itself.  Generally, this is not considered to be best
practice, however there are some situations where it is the best
option, so it is permitted on an opt-in basis.

When enabled, there are 3 ways to access the feature.  All of them
return the pointer reference held by the container.

:: From the Object/Mix-in ::
Mix-ins expose a RemoveFromContainer method which may be called.
Objects which derive from one of these default mix-ins may simply say:

   my_object.RemoveFromContainer();

:: Using the global fbl::RemoveFromContainer ::
When objects are using the ContainableBaseClasses pattern to exist in
multiple containers simultaneouly, the easiest way to find and invoke
RemoveFromContainer on the mix-in is to use the global helper.

  fbl::RemoveFromContainer<ContainerTag>(my_object);
  fbl::RemoveFromContainer(my_object);  // DefaultObjectTag

:: Directly from the node instance itself ::
If a user is in a position where the must use custom node traits to
find their node within the object, they may still use
RemoveFromContainer, but it will require the NodeTraits used to find
the node in the object.

  struct MyNodeTraits { auto& node_state(Obj&) { ... } }
  my_object_.my_node_.template RemoveFromContainer<MyNodeTraits>()

:: Restrictions ::

RemoveFromContainer currently only works with DoublyLinkedLists which
do not internally track size.  Maintaining the internal size
bookkeeping in O(1) is not possible unless the user knows where that
bookkeeping is (IOW - they have access to the container).

SinglyLinkedLists will never be able to implement this feature, by
nature of the data structure.  Likewise, HashTables which use SLL
buckets will never be able to do this either.

This leaves HashTables with DLL buckets and WAVLTrees.  Currently,
neither of these support RemoveFromContainer because they both track
size internally.  If there was a good use case, however, this
restriction could be relaxed by introducing versions of the containers
which did not track size.  While this would be trivial for HashTable,
it would be more complicated for WAVLTree.

Tests: new unit tests.
Original-Fixed: 49965

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386830
Original-Revision: 9cdf1bdff84bf6a63d988d498ee59b24997f9033
GitOrigin-RevId: 2c9ba2806181f0c86b0ad17f79fda884acfda423
Change-Id: I65747f53a994a7faa7933dbee5cc54a641621dd0