[roll] Roll fuchsia [power] Components can state why the power state is changing
This change is part of a larger effort to expose more granular
information on why device's reboot gracefully. It adds a method to the
Admin protocol that allows callers to specify why they are requesting
a change in the device's power state. Additionally, this change adds a
protocol that allows components to watch for requests to change the
system's power state by registering a callback with the server of
Admin.

Original-Bug: 52190

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390219
Original-Revision: 1d7879fb8463d4f2aee47918b61dc4af8866b04b
GitOrigin-RevId: c817685350ed7d1c0cb3cfaf171c9dd10db23b53
Change-Id: I51acfeb74bea46c795ab303228e5dcb6ea1bc843