[roll] Roll fuchsia [intl] Forbid comments before document start
Today, I learned that it is not up to XML spec to have
a <!-- comment --> before <?xml?>.  Since xml-rs parser
doesn't enforce this constraint by default, we needed to do that
ourselves.

This allowed us to let non-well-formed XML files through, and fail late
in the process.   Since we should fail fast, I modified the parser to be
fussy.  And since we must verify that the top-level programs inherited
the behavior (since behavior is a parser config), needed to add tests
throughout.

Reported a bug as https://github.com/netvl/xml-rs/issues/193, see there
also for details why this is wrong.

Original-Fixed: 52339
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390661
Original-Revision: c91a35c839737090676458ef69086b44d8591954
GitOrigin-RevId: 55739c23fe7266fa34b17c11cae8db87350eacb0
Change-Id: I99926b5152f23e8f9c7e5580761058bc527f9f0b