[roll] Roll fuchsia [bt][avrcp-target] Return 0 if reference time is in the future.
It is valid for a client to provide a TimelineFunction in which the
reference_time is in the future. However, when calculating the
song_position, this can sometimes lead to negative song positions. Since
AVRCP only relays non-negative song positions, clamp the calculation to
zero.

Removed the warning that was issued if the reference_time was in the
future (i.e the diff was negative).

Original-Fixed: 52318
Original-Bug: 52254
Test: Updated
bt-avrcp-target-unittests: test_timeline_fn_to_song_position
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390759
Original-Revision: abe707afdcb26ea37c7c6079e58fa179edc56598
GitOrigin-RevId: 317de31486f0309b12326a8dbb5b3967e1d8e989
Change-Id: If59d3cfd7ab77e27a81b7975c4a3b2d7214f81ef