[roll] Roll fuchsia [async] Move async-default to its own folder.
This will allow it to get migrated to the GN build while the rest of async
is currently blocked by FIDL dependencies.

This change shedded some light on a few places in the codebase where the library
was not properly depended on.

Original-Bug: 36548
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390786
Original-Revision: 69643b066e57628428ace92a78dd6757f738991f
GitOrigin-RevId: c40a9ce2b11b5896a9f60eb81f6ac21ed78adcbd
Change-Id: I013039c142854d68feedd89926cc941ff16f2a90