[roll] Roll fuchsia [flatland] ReleaseImage implementation.
The primary purpose of ReleaseImage() is to prevent future Flatland
calls from using the ImageId and make the ImageId available for reuse
in a future CreateImage() call, at which point it will be associated
with a different Image.

Calling ReleaseImage() does not remove the Image from any Transforms
that it is currently attached to, nor does it prevent it from being
rendered in the current or future frames.

The Image can be fully garbage collected once all references to it in
the Transform hierarchy have been deleted (either by calling
SetImageOnTransform() with a 0 image ID or deleting the Transform
altogether). Until then, the ImageMetadata is kept alive by Flatland.

This change does not include logic for cleaning up cached image
resources in the Renderer.

TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386849
Original-Revision: 1f8b143e062064d999e0c57a011588a1d5a35498
GitOrigin-RevId: 29191af04ff9136c994ff20f833ce2b1c7432574
Change-Id: I9ab71577ad5993497131d6eeafea81a68803a347