[roll] Roll fuchsia [wlan][policy] Remove network
Adds the 'remove a saved network' functionality for the policy client
API. The credential and other data about the network will be forgotten,
so in order to auto-connect or manually connect through the client
provider API, it will need to be saved again.

Removing a network will cause a disconnect if we are currently connected
to the network through the new API. If we connected to the network
through the legacy code (including current auto-connection behavior),
removing the network will not trigger a disconnect.

Tests: added/modified unit tests
	 fx test wlancfg-tests
       manually tested remove using donut CLI tool and save/get,
       including disconnect when removed and persistence
Original-Bug: 49581
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378513
Original-Revision: 38e4588db85055f68f9757b1dcc6cd3f0a4ac00a
GitOrigin-RevId: 5f415d77f969b0e355529adc74346532243b30d1
Change-Id: I059d7f9fb326e6a53440fa3d10bd48e8414b7f4d