[roll] Roll fuchsia [fxtest] Make output_test complete sucessfully
Previously, FuchsiaTestCommand would set the global exit code when it received
a TestResult. This worked correctly in production, and it worked when run in
automated tests with pub run test. However, it would cause spurious test
failures when run using fx test (i.e. a GN-based build). This caused
output_test to fail with exit code 2 despite all the tests in the suite
succeeding.

This difference occurs because of the way Dart tests are packaged with the
dart_test rule. In pub run test, the test process doesn't affect the exit code
of the Dart test runner. When packaged by GN, the test itself can set the exit
code of the overall process as long as this occurs during test suite (group?)
teardown.

That said, tests should not modify global state. The exit code variable is a
global and should not be modified by the test or the code under test. This
change replaces the direct modification of exitCode by a function that can be
safely replaced in tests.

Original-Bug: 52244
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390479
Original-Revision: 777e9cc50c234926b9279980e7082b475b2fc315
GitOrigin-RevId: fa8d235214f33a3fd7fe316e52c9abe2338b3532
Change-Id: If317759a4881d133608230278f079657dcce2e77