[roll] Roll fuchsia [isolated-ota] Support using omaha to install updates
In some instances we want to use Omaha to determine the update package
to install, along with the URL of the update package. When isolated-ota
is supplied an Omaha app id, make it use Omaha to get the update
metadata.

This CL adds a custom Omaha installer for isolated-ota which leverages
the existing sandboxed system-updater implementation to perform the
installation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386918
Original-Revision: 5f741535d44ad7130c739203612eb47cd5704b3f
GitOrigin-RevId: 57c4c5d10efa1e7178df9a8343b2d7a2dd2b5ac9
Change-Id: I8f2b7e28f8a13e7718d72adbe673783ca18d7243