[roll] Roll fuchsia [kernel] Move ownership of root job to Executor.
Currently the RootJobObserver is responsible for creation of the root
job, and has the only reference to it. Methods in Executor pull out
the root job dispatcher on demand.

Move responsibility for ownership / creation of the root job to the
Executor class, passing it in to the RootJobObserver as an argument.

Original-Bug: 51463
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390974
Original-Revision: a41148e9bc6f8440b42f400387a539100e424576
GitOrigin-RevId: 182804564dbbc23a48d59bf78d48cc43ce7102d1
Change-Id: I76e4f67a3d4bfe595948c98244f7ccb67476bbdb