[roll] Roll external/github.com/catapult-project/catapult auto-triage: File issues on specific projects
This change implements support in the Dashboard for auto-triage to file
issues that match subscriptions defining multiple projects. The change
involves two major pieces:

1. We're creating one group per Monorail project for matching
auto-triage configurations. This allows us to differentiate alert groups
for different projects which match the subscription configuration. This
allows folks from different projects to file individual issues for the
same alert group for tracking the fixes/investigations independently.

2. We're preparing the Anomaly model to support multiple issues with
projects associated in the model. This is only in preparation until we
find a way of representing this in the Dashboard UI, and until we can
ensure that we can transactionally update the list of issues to avoid
eventual consistency conflicts.

We're also taking some opportunity to refactor the fake issue tracker so
that it has value/object semantics instead of being a global.

R=abennetts@google.com, fancl@chromium.org

Original-Bug: chromium:1082032
Original-Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2206622
Auto-Submit: Dean Berris <dberris@chromium.org>
Original-Revision: f3dc334579056f6212ace5db04ae01fe72277f9e
GitOrigin-RevId: 1e3a1d28cd087c6b5d841aa87b42ff8a75a048bc
Change-Id: Ic87564a6e8c87f5d3946e090716e00a5c4129e45