[roll] Roll fuchsia [aml][cpu] Add aml CPU as a child of the composite device
AML CPU wasn't being added as a child of the composite device
which was causing the driver manager to complain about a failure to
add a device in bind.

Tested by observing that the "failed to add device in bind" message
no longer appears in the logs and that the cpuctl functionality still
works.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382913
Original-Revision: 32f0210e560bdcd3cb430499e912d63f3824db77
GitOrigin-RevId: 68ae5e8d4581f45620219a29cf70d01fda0548e8
Change-Id: I3822db2cc9da034232f2f2c8506e6b1eddb4299e