[roll] Roll fuchsia [logs] Define preliminary severity encoding
This defines a field for encoding severity on
the structured log header. This is a preliminary
location and may be moved later on - this is
acceptable for now as only internal diagnostics
libraries use the encoding.
This change also updates the severities to the
new definitions.

Testing: fx build

Original-Bug: 50517, 50106

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391192
Original-Revision: 1dc12f1d3fc02fe3ac52a05efea43af1fa9eb6c9
GitOrigin-RevId: ce4d3a0a6ea4a6a08a2e365d46e0c6a9629d2d6d
Change-Id: I0c1a6f50f62fa91aed0d628167cfddf459254c4a