[roll] Roll fuchsia [blobfs] Make compression level configurable
Adds --compression-level flag, which can be used to configure the
compression level blobfs uses for newly written blobs.

The flag is only respected for ZSTD* variants. A runtime error is raised
if the flag is set to an invalid value, or if it is set when a different
compression level is used.

Original-Bug: 52261
Tested: Booted; Unit tests.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391364
Original-Revision: 18526cebe466664c1dd326476c46208af50e9c01
GitOrigin-RevId: 4d81338363c875c157a50c8ed12864176db1d477
Change-Id: I7d5c993ef8b9808b2e9cd269203b2ede265b7d96