[roll] Roll fuchsia [debugger] Handle line 0 stepping better.
A previous change to avoid reporting "line 0" entries in stack frames
caused stepping to stop at these entries. This change clarifies this
behavior in comments in ModuleSymbols, and fixes the behavior.

The solution is to remove the check that the stack frame's file/line
match the line entry for the given address. This was added to try to
detect the case where the entry was at the beginning of an inline frame.
but we can check this condition directly instead, and use the
information from the line table consistently.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390776
Original-Revision: 22e3bbdbff54793c54787d968d3f18330b296dc4
GitOrigin-RevId: 93a4008c32822985ef4655a45cd158943b0c81b5
Change-Id: Icc58d27ec4972c2c612626b371b9eebacee169b6