[roll] Roll fuchsia [fidl][llcpp] Fix Close() path in Reply().
In AsyncTransaction::Reply(), errors cause the binding to be Close()d.
This can occur at a time when the AsyncTransaction has ownership of the
binding's internal self-reference. Previously, that self-reference was
not being returned to the binding prior to the Close(), potentially
leading to crashes. This CL fixes this behavior to restore the reference
and ensure that no invariants are broken.

Original-Bug: 50822,42086
Test: runtests -t fidl-async-test-test -r 100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390639
Original-Revision: 41ac114f936829bcfdaa4c07a83fa112ed9a475e
GitOrigin-RevId: 8f26b7b542aede2442f6576dcc75ecd33486d328
Change-Id: I812d5284d4f0f7dcdb69c9fb639632af3fc648b9