[roll] Roll fuchsia [media][audio] Fix flake in audio-loopback-tests
The audio-loopback-tests timed out when searching for the expected
packet in the loopback capture buffer. This is due to the packet data
being offset in the capture buffer beyond the expected payload offset.

To address this issue, the test now searches the capture buffer for the
true offset of the packet data, awaits all frames to be captured, and
then checks the capture buffer for the expected data, starting from the
true offset.

Multiply: audio-loopback-tests

Original-Bug: 47968

Test: fx test audio-loopback-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377074
Original-Revision: 398542dff4f564feca7898989c7c10c5966f24b7
GitOrigin-RevId: 193971bed1d5df5588e92708b102a063f0c9b8ed
Change-Id: Ie7c19a6e3c217853fc23c536f93297de9d0274d8