[roll] Roll fuchsia [fshost] Remove dependency on hermetic compute
This CL is part of a series of CLs that migrate the system away from the
hermetic compute library. Currently, the cost of maintaining hermetic
compute outweighs the benefits. Rather than decompressing the ZBI
out-of-process, we now decompress the ZBI in-process in fshost.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391419
Original-Revision: 2a25ecd46bdff2b99a67b868b14c9d1caa8439dd
GitOrigin-RevId: f5c7eb2abf02a9da55f774172fc47b0b70e4d581
Change-Id: I24379a6d2a96c6d708d1f7b123bfdcba6f4bfbea