[roll] Roll fuchsia [gidl] Generate multiple output files in GIDL for C++ targets
Currently, GIDL is generating a C++ file that takes a long time to
compile. This change splits up the file.

Originally, I hoped to name each output file based on the type name and
generate a GN dependency on that file. Unfortunately, this is quite
difficult to do in GN. For expediency, GIDL now generates a fixed number
of outputs, leaving unused output as empty files. This means specifying
the fixed number of outputs to be a larger number than the actual
generated outputs. However, if this number is exceeded, it is easy to
add additional outputs.

Original-Bug: 52131

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390528
Original-Revision: 992305ee6819e43b74958dee41d6ca451608d0ee
GitOrigin-RevId: d4a864b2ae26613a9260be44f270d17355298000
Change-Id: I88c2233fa0cb10cb090be7dc5e504a07fc4fb68e