[roll] Roll fuchsia [fbl][slab-alloc] Add AllowManualDeleteOperator option.
Rework how template options for slab allocators work, similar to what
was recently done for fbl:: container NodeOptions.  This rework adopts
an enum class flag pattern along with some operators defined for the
flags which make them easier to handle.  The result is the ability to
have more a strongly typed set options which end up reading in a
descriptive fashion instead of simply being "true" or "false" in code.
Additionally, it makes it a lot easier to add new flags at a later
date without needing to add and manage new template parameters.
Finally, it also moves away from the style of ALL_CAPS_CONSTANTS and
in the direction of the preferred CamelCase::Option style.

Finally, define a new options which skips the DEBUG_ASSERT present in
the destructor of slab allocated objects which belong to a manual slab
allocator.  This gives users of manual slab allocators the ability to
define objects which may or may not be slab allocated.  Users must be
very careful to keep their own bookkeeping clean, and return the
object to the heap or to an appropriate slab allocator, but this is an
advanced option sitting behind an opt in flag, so I feel OK about
that.

Tests: new unit test, and some manual negative compilation tests.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390793
Original-Revision: 512111c9627c83981a48b78627e3095cf083c239
GitOrigin-RevId: 37b5334ddbdabe6d6c55013e1fd08d7e755e2712
Change-Id: Ia9d7a8c414babd4e51f89c9553a91b88e723e96f