[roll] Roll fuchsia [kernle][vm] Unify VMO Commit and Pin operations
Performing Pin as a separate operation on VMOs introduces the chance
for race conditions to cause the zx_bit_pin syscall to fail due to a
decommit happening between the Commit and the Pin.

Although Pin is allowed to have transient failures, doing so
unnecessarily places burden on clients and is less efficient overall.
In this case, preventing a race between Commit and Pin is a minimal
increase in complexity for the kernel, whilst greatly increasing the
Pin success rate to clients.

The included test case is not perfectly realistic since if clients
are Decommitting their own VMOs then they probably shouldn't epxect
to also be able to pin it. Howver, the test simulates the idea of
the kernel acting as the decommitter due to eviction, zero page
deduplication etc.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388093
Original-Revision: 2b3b9445ee73585f4cbf8c2fd0abc8067c32ed08
GitOrigin-RevId: aa2b4b72fca4ee36b784c295d2e0f117073b139a
Change-Id: I18ccca50b6dd7099df631f1628fa8662f54c3720