[roll] Roll fuchsia [virtualization] Gates calls to connect to vsock.
server_daemon_test does not wait for a vsock connection to
be made before attempting to use it.  If the socket is used
before it is ready, writing to it fails resulting in a test
flake.  This change waits for the call to connect to the
vsock to complete before continuing with the test.

In testing this patch, it was discovered that the
guest_interaction_daemon being alive does not guarantee
that the program is listening on the vsock.  If the daemon
is not listening on the vsock, the call to connect will
fail.  This patch adds a log after the call to listen.
Tests can then query journalctl over the guest's serial
console to determine the state of the daemon before
attempting to connect.

Original-Bug: 36321
Test: fx run-test guest_interaction_tests -t \
      server_daemon_test
      fx run-test netemul_guest_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/365433
Original-Revision: 28e09c0482168b2fd935ea02a1da1c420f6e6c7d
GitOrigin-RevId: 1312a71b7a364801446be91263f603f2a07b6dd8
Change-Id: Ifeb60651b29affaa2cc4009d69500e5d2a364e27