[roll] Roll fuchsia [fidl][llcpp] Allow dispatcher shutdown.
This CL removes the assertions that the dispatcher is not being
shutdown. The trade-off is that on shutdown, the unbound handlers may
end up running in the thread calling shutdown. The user must ensure that
no locks are held around it (though they should probably be doing that
anyway since it blocks on joining dispatcher threads).

Original-Bug: None
Test: runtests -t fidl-async-test-test -r 100; runtests -t fidl-test -r 10
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391274
Original-Revision: db8e7bde8551c23373b23b03ab5ba421ba889df6
GitOrigin-RevId: ad67610ee8428a109ab05e53bb6a74fd618b507f
Change-Id: I29ec56343c03e8ec2d487788e3c23d41f9e6f69d