[roll] Roll fuchsia [audio_core] Driver owns reference clock
The AudioDriver object creates and maintains a zx::clock
object, representing the hardware device clock. It will
actively adjust this zx::clock object to reflect the actual
hardware rate that it detects. It also creates a read-only
duplicate of this clock, to be read by other entities
throughout the audio pipeline. A reference of this read-only
duplicate is passed to the AudioDevice, who will forward it
onward as needed into the OutputPipeline or capture path.

Original-Bug: 46625
Test: new cases in audio_core_unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385254
Original-Revision: d3131e450a6e619e29e4e157ae3851fb7ed9ce05
GitOrigin-RevId: aaf8c541445b3ffe2250d0c5b37ea4409d58a670
Change-Id: I2b19f979f9353dc4b50e39dca98fcb4a632db64c