[roll] Roll fuchsia [storage] Update blobfs metrics.
These metrics are only printed out at filesystem unmount so this should
not affect any server metrics.

Removes counting the pre-cache of blobs as a "lookup." Originally the
lookup metric tracked blobs as they were opened. Later a pre-caching
layer was added in Blobfs::InitializeVnodes() and this metric was added
there.

Combining both a count of all blobs in the system plus all blobs opened
doesn't make sense. This removes the lookup metric counting from the
pre-caching so that the lookup metric counts only blobs accessed via
Directory::Lookup. If we need more filesystem-wide blobfs metrics, those
should be added in a separate value.

Removes the "writeback thread" metrics. These were never updated and
reflected the design before there was a journal. If we want this type of
data it would be a journal statistic we would need to add in separately.

Timings are only taken when Cobalt metrics are enabled. Since otherwise
the only use of these statistics is a printout to the console that users
will never read, this behavior is preserved. To avoid being confusing,
printing timings are disabled to avoid confusion.

Original-Bug: 52351

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391496
Original-Revision: 40db081db0d546c3b90054b07420a4ae4c2f527a
GitOrigin-RevId: 5608af6acc4ee25d8a809d05a8e6035f1af816b8
Change-Id: I5c0049041f3e358210c68ac5715d864ea26970be