[roll] Roll fuchsia [usb-device] Verify wTotalLength sanity
A malicious USB device could change the value it returns in wTotalLength
when the USB host reads configuration descriptors, because those
descriptors must be requested twice -- once to read only the header of
the descriptor to learn the length, and then once again to retrieve the
entire descriptor into an appropriately-sized buffer.

We already verified that the number of bytes we read back in the second
request matched the length specified in the first, but we did not verify
that the contents of the second request matched those of the first.
Failing to do so would leave a descriptor that claims to be longer than
the buffer allocated to it is, which if not handled very carefully could
lead to out-of-bounds reads.  Indeed, the rest of the code treats
wTotalLength as authoritative.

If a device attempts such trickery, we should reject the device in the
same way we'd reject it if it gave us a short read.

Additionally, we should reject wTotalLength values that are shorter than
the config descriptor header -- they can't possibly be valid.

Test: added new test for this case to usb-device-test that fails before
the change but passes after.

We're grateful to Quarkslab for reporting this vulnerability.

Original-Fixed: 50619
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391569
Original-Revision: fd872ce1d485f9a2ee2793aa2e113f84352d5794
GitOrigin-RevId: 7bee934091ca89386221faf7af31eba190ae783b
Change-Id: I4a71f77276f688879d08d7386247db45b2f4e5ad