[roll] Roll fuchsia [sl4f][process_trace] Fix build error for process_trace
The build rule makes an implicit library for the binary and was using
the same directory as the root for both the sl4f client library and the
implicit library, causing conflicts. This change results in the root
changing from client/ to client/bin/, avoiding the conflict.

TEST: Include //sdk/testing/sl4f/client:process_trace and fx build
BUG: 52435
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391499
Original-Revision: 7eb5530eefdfbb60aef35335562bb372476c74be
GitOrigin-RevId: bf42af486d816141289cdfb88e6d2ada4d69ea7d
Change-Id: Ia274fe7c3c707d4f84d28aef2294daaf3612385b