[roll] Roll fuchsia [audio_core] Flush queue on unlink
Fix a use-after-free error condition that can occur if there is a
mix-job in progress when a renderer is removed from the system. When
this happens, the outputs mix thread may be holding a reference to
in-flight packets while RouteGraph is performing unlinks and removal
of the Renderer. In some situations, the Renderer can be free'd while
the mix job is still retaining packets (and a packet queue) from that
Renderer.

This is problematic because we use a SlabAllocator, which is owned by
the Renderer, to allocate those Packets. So when this happens we leave
pointers to invalid memory in the mixer thread.

We resolve this issue by running a Flush on the link's PacketQueue
during unlinking and passing a strong reference to the Renderer to the
Flush callback.

Original-Fixed: 49622
Test: fx test audio_core_unittests
      New test fails with an ASAN violation without the patch to
      base_renderer.cc
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391595
Original-Revision: f946b89b03c2270c81b4abd5581d46185a13e641
GitOrigin-RevId: a4528f9c8a8e596ebdf93e048a19295101c48870
Change-Id: Ie7df32cc9899f50598f17e74279e64bc924d6ae9