[roll] Roll fuchsia [kernel] Hard-code the "count" field of signal port packets to 1.
Currently, port packets created in response to the ZX_CHANNEL_READABLE
signal firing on a channel object will fill the "count" field of the
packet to the number of pending messages.

This CL hard-codes this number "1" in preparation for the field's
deprecation. This changes the semantics slightly:

  * For channels, instead of containing the number of messages at the
    time the packet was enqueued, the field will simply be "1". Message
    loops, instead of reading "N" messages off the channel in a batch,
    will instead read a single message before setting up a new
    zx_object_wait_async request.

  * For other packet types, the value of "count" will be "1" instead
    of "0".

Forcing userspace code to read one message at a time instead of batching
initially seems like a loss. In practice, however, observed workloads
show that ~98.4% of channel port packets have a count of "1", and ~93% of
messages are in a batch of size 1. The change will result in a 1.6%
increase in zx_object_wait_async syscalls for these workloads.

In return, these changes allow simplifications in the kernel code in
upcoming CLs which will give a performance boost to the common case.

If workloads change in the future and batch sizes increase, one option
is to modify the "zx_channel_read" syscall to indicate if more messages
are still available on the channel, allowing us to keep these
simplifications while avoiding the extra zx_object_wait_async set up
overheads.

For other packet types, we can audit the current in-tree code to find
all users of the "count" field:

  Message dispatch libraries:
  * sdk/lib/fidl/cpp/internal/message_reader.cc
  * zircon/system/ulib/dispatcher-pool/dispatcher-channel.cc
  * zircon/system/ulib/fidl-async-2/include/lib/fidl-async-2/simple_binding.h
  * zircon/system/ulib/fidl-async/bind.c
  * zircon/system/ulib/fidl-async/llcpp_bind.cc
  * zircon/system/ulib/fidl/llcpp_async_binding.cc

  Message dispatch code:
  * src/connectivity/bluetooth/core/bt-host/hci/acl_data_channel.cc
  * src/connectivity/bluetooth/core/bt-host/hci/command_channel.cc
  * src/connectivity/bluetooth/tools/bt-intel-tool/command_channel.cc
  * src/connectivity/wlan/drivers/wlan/device.cc
  * zircon/system/ulib/logger/logger.cc
  * zircon/system/ulib/runtests-utils/log-exporter.cc

  Marshalling / ABI validation:
  * zircon/kernel/lib/abi_type_validator/include/lib/abi_type_validator.h

  Tests:
  * src/lib/fidl/llcpp/tests/flexible_test.cc
  * tools/fidlcat/interception_tests/port_test.cc
  * tools/fidlcat/lib/syscall_definition.cc
  * zircon/system/utest/core/port/ports.cc

Outside of tests and ABI validation, all uses of the count field are the
"message loop" case. In particular, no in-tree code is currently looking
at the "count" field for non-channel packets.

Original-Bug: 50601

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377457
Original-Revision: f197744e67a4413cc350ac2ee6d4a174c02a606b
GitOrigin-RevId: 72d8c257d5a44fd49e9545c3d6bd4cd6ee528330
Change-Id: Ia6da9ecd1db74d4a219c717bae1b76e9a8162ee4