[roll] Roll fuchsia [i18n] Adds a time correctness system test
This test will fail if the Dart SDK configuration changes such that the
dart timezone is broken on a running device.  It should prevent the "OMG
the clock is showing the wrong time" which sometimes happened -- or, at
least, it will prevent such cases where there is a discrepancy between
the system timezone and the runner timezone.

This test has been a long time in the making.  In order for this test to
be able to pass on an actual device, the following changes were needed
(some by me some by others):

- Adding a global component context to dart VM.
- Adding dart VM inspect
- Fixing the basemgr wire-up of locale context bug, which caused the
  locale change to be invisible in the dart VM.
- Fixing the Fuchsia flutter roller.
- Fixing a certain misplaced `assert` statement.

Original-Bug: 43341
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/367445
Original-Revision: f31ff1423eacf5e0d3f7fa288335689a65c3ba17
GitOrigin-RevId: 591599450a2ad5aeffd32922bcc80981a45e272d
Change-Id: Ib9b5d5be64996c82486af5eb6b9bb1cbf4bc84ee