[roll] Roll fuchsia [media][audio_core] Enable input driver_gain_db
On driver initialization, driver_gain_db is applied to the input driver.
The driver_gain_db value is sourced from the device config; if the key
is missing from the config or the config is not available, a default
gain value of 0.0db is applied.

Original-Bug: 51003

Test: fx test audio_device_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391422
Original-Revision: 7f273d362b13afa7d47a04cf80795d70b059a425
GitOrigin-RevId: 5432f07fc2a8f92d7b65a86a9dbb051ea3d03a60
Change-Id: I49595789a69d4c7c8a624f97a90ba47160c30665