[roll] Roll fuchsia [inspect][rust] Add weak node references.
Weak references is relevant primarily in cases where nodes or
properties are added post-instantiation. For instance, consider a
type that has an `Option<StringProperty>` that is attached to the
parent. If initialized to `None` upon construction (when parent
is usually provided), there is no way to attach to inspect at
a later point, without providing parent in the function chain on
non-constructor regular method calls.

For `fuchsia_derive_inspect`, this unlocks having `Option<T>` be
represented as a node or property that is absent when `None`.

This change moves the custom drop-logic to the innermost type,
the `InnerRef`. Since deallocation depends on the node or property
type, a new trait `Free` with different implementations are provided.

Original-Fixed: 51720
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390853
Original-Revision: 4b97ee78fa7858e7bc7967c00169d3ac33994659
GitOrigin-RevId: 65cf5239072e68cc552830024a7c0137b2751d4f
Change-Id: I25d62dba403262e705adf15250bb835ed75b99e1