[roll] Roll fuchsia [inspect] Refactor NodeHierarchy serialization
This first pass implements Serialize on NodeHierarchy directly.
The reason it was done separately is purely historical when
NodeHierarchy used to be in a different crate.

This also makes the `Schema` in Archvist generic and not
depending on the format. This simplifies introducing new formats such
as CBOR, FIDL Text, IQuery Text, etc.

Additionally, the Schema is now
generic over the Key. This makes it easily extensible to support Logs
that have a different Key type than Inspect and Lifecycle.

Unfortunately, deserialization is a bit more complicated given that
it's over-fitted to json at the moment and simply manually implementing
Deserialize is by far not as straightforward as implementing Serialize.
A follow up will refactor it so that we have
`impl Deserialize for NodeHierarchy`.

Tested: just a refactor, existing tests continue to pass

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391560
Original-Revision: c344db12ec84030b08bba39672c127fd161491fd
GitOrigin-RevId: 2712b21eb0c32e59b113a7a3ad1f6362b3bf6fdc
Change-Id: I8c9925642acc2d065a254b942a6f4421f678b389