[roll] Roll fuchsia [region-alloc] Relax RegionPool rules.
Previously, to use a RegionAllocator, you were required to use a
RegionAllocator::RegionPool object as the SlabAllocator based source
of bookkeeping objects.  After this change, these rules are relaxed.
A RegionAllocator may still be used with a Pool, but it can also
operate without one, simply allocating from the heap.

This is preparation for a new use in the kernel where the allocator
will be used mostly for testing for attempts to create resources and
MMIO mappings which could conflict with a set of forbidden ranges.
The RegionAllocator provides a convenient (and well tested) API for
doing stuff like this, but using a RegionPool is way overkill in this
case.  The forbidden regions will be set up once at startup, and then
never changed.  Minimizing heap fragmentation because of bookkeeping
churn is not a concern in this case, and potential overallocation
because of slab allocation is undesirable.

Tests: new unit tests.
Original-Bug: 34029

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390795
Original-Revision: ca79b838ea61b5c31fd45a302ad3ee5e788314e6
GitOrigin-RevId: 7c0707a9a43eb81e8de7d6e8d9466fe52686126a
Change-Id: I60ce3106405036866351cd01dec9023efa5375ce