[roll] Roll fuchsia [modular] Fix flake in new agent_runner_unittest
When the AgentController channel is closed, it is similar to sending a
final message on the channel. Before a Binding or InterfacePtr will
return !is_bound(), the run loop whose dispatcher is waiting on that channel must have a
chance to consume and process the close message.

Hypothesis: There were two tests that were not waiting for this to happen, checking
if is_bound() was false based on a synchronization point that could
arrive before the close messages did.

I could not reproduce the original flake with n=10000 retries. This
change still works and the tests still do not flake with n=10000
retries. I suspect this will fix the underlying race condition.

Original-Fixed: 49609
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392237
Original-Revision: e0d605ea84e83da846263486b0e6307b6f558b78
GitOrigin-RevId: ff67ca57abecd3304bb9fdf642a7aeba8961bf86
Change-Id: I49604ad64b9dffbb94ee3291ed241ecc4d23934b