[roll] Roll fuchsia [isolated_devmgr] Catch exceptions when driver_manager crashes.
We were catching these exceptions in the component library built on top
of devmgr_integration_test, but that is only used for tests which spawn
isolated devmgr as a separate component. This patch adds support for
tests which spawn the isolated devmgr directly.

Original-Bug: 51679
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390525
Original-Revision: ba4745401cbcacc16ce0d1cd4252e64f812597ce
GitOrigin-RevId: 568f4ca35c0193412e61eddfc967c09a2b8fe037
Change-Id: I3257d28afb4f576ae62fc2fd05afcde46fdc66b3