[roll] Roll fuchsia [shell] implement shutdown
This allows the client to destroy the interpreter it created. When the
interpreter is destroyed, it checks that everything is deallocated
correctly (there is no memory leak).

Fix a memory leak for String when the string was created using a
StringContainer.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390585
Original-Revision: eaa8d7498107cb1293c81ba7d6f9701dd482712a
GitOrigin-RevId: 7a8717aeae5ad6b5f32e27ad47c132baae78824c
Change-Id: Icebaab9bb002bc321bc55241715d6d6b2289c66b