[roll] Roll fuchsia [trace_processing] Re-add ("gfx", "RenderFrame") metric
Re-add a metric for ("gfx", "RenderFrame") durations.  Given that
there's an active regression for this metric, we'll want it to verify a
fix for the regression.  It also might still be useful to track as a
submetric.

Test: ./out/default/sl4f_client_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392325
Original-Revision: 6da03955e2e230e198d3aea1bc8e92d65ce6ce9a
GitOrigin-RevId: 016341e210360f88dd30ce9db5dbaf2b7a92226b
Change-Id: Ie07075040be531640f97a4f2da99501a79a40145