[roll] Roll fuchsia [location][fidl] increase precision of position
Summary
-------
The current definition of `Position` represents the `latitude` and
`longitude` as 32-bit floats, which is insufficiently precise.

To resolve this, we change those fields to be 64-bit floats.
While there, we change the `accuracy_meters` and `altitude_meters`
fields to be 64-bit as well.

Note that we probably don't need the additional precision for
`accuracy_meters`, but it seems simpler to use the same type
for all the fields.

In detail
---------
With 32-bit floats, the mantissa is 24-bits, allowing us to
precisely represent values in range of 0-2^24-1 (the latter
value is approximately 16.8 million).

One degree of longitude at the equator equals approximately
111 kilometers [1], and longitude ranges between -180 degrees
and 180 degrees.

To support 1 meter precision, we need to precisely represent
values down to 1 millionth of a degree. (1 degree / 111 km,
rounded to the next order of magnitude.)

Given that the values range up to 180 degrees, we need to be
able to represent 180 million positive values precisely.
However, 32-bit floats can only precisely 16.8 million
positive values.

Hence, 32-bit floats are not sufficient.

[1] https://www.thoughtco.com/degree-of-latitude-and-longitude-distance-4070616

Original-Bug: 51513
Test: fx build
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391793
Original-Revision: c9c7da3fafe480b286916f5605f09d62d027298f
GitOrigin-RevId: a3998300536fbc6021530c0f466ba861c25041d5
Change-Id: Ia7f7691d8312282af8eec2373d52d1c1195ffe7c