[roll] Roll fuchsia [region-alloc] Add intersection/containment test operation
Add the ability to test a user defined region against the set of
either allocated regions, or available regions, to see if the user's
region either intersects any of the regions in the chosen set, or is
completely contained by any region in the chosen set.

This makes it easy to use the region allocator as type of region
allow/deny list.  We can deny a region if it intersects any of the
regions in a forbidden set, or we can allow a region if it is
completely contained by any of the regions in the allowed set.

Tests: new unit tests
Original-Bug: 34029
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390796
Original-Revision: 12c049e6fc1f05f25df0f05063add87e143a44ac
GitOrigin-RevId: 4d1bd5cc59f17a3e095f8a46f71ab7d050c77c1f
Change-Id: Ibdd2430a711e29cefa35ad18cce8d5000e36069c