[roll] Roll fuchsia [intl] Fixes a flake in intl_services_test
While I'm not certain of the exact mechanism, the flake was caused by
the binding set being empty on the server side at the time the
`OnChange` event was supposed to be delivered.  This meant that the
test never received its `OnChange` notification so never progressed
beyond the point where it waits for it to arrive.

I suspect the reason for this was that we built a callback setup first,
then called `RunLoopUntil` which processes the binding to the server
among other things.  And if the binding happened to be processed after
the OnChange was supposd to be sent, we miss an OnChange event.

Resolved by introducing a dummy call to GetProfile at the very beginning
of the tests, which ensures that the server is aware of a client's
existence, before we start the "real" tests.

The problem was rare in regular execution, but by attaching `fx fidlcat`
to the test as shown below it became almost 100% reproducible. I guess
the watched programs became slow enough that the data race became
prominent.

```
fx fidlcat --remote=name="intl*"
```

Multiply: intl_services_tests: 80
Original-Fixed: 50789
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392776
Original-Revision: ac552e44bdd207594d4a0912413e369bdcf32c1e
GitOrigin-RevId: cfa07ff8dc1c90019ed92eec2d4363840445bab9
Change-Id: I619bee121b0473344336c0824ff3bb9a5897d021