[roll] Roll fuchsia [modular] Remove Login Override.
Login override was a holdover from early 2019 when we had a base shell
that handled user signin. Its been unused since ~Q3 last year and now
that we've removed the concept of a base shell there wouldn't be a way
to resurrect it. This also removes the basemgr dependency on SetUI.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392295
Original-Revision: d3c1ccb2c2ddbcca330d4cd4a467f4f67229497a
GitOrigin-RevId: 4a56855ca98eb7ecd9a1e52523a05bad9a47d359
Change-Id: I12730386e5ed3e316b0704ae658576abd03280a2