[roll] Roll fuchsia [fidl-lint] Enable color output in fidl-lint
Currently, fidl-lint does not explicitly enable color output in any
case. However, it was sometimes erroneously enabling it due to a bug.
fidl::reporter::Format takes a bool `enable_color` and also a parameter
`squiggle_size` with a default value. fidl::utils::FormatFindings was
passing in `squiggle_size` for `enable_color`, which silently compiled
because it is a size_t reinterpreted as a bool, and because the next
param has a default value.

This CL fixes that bug and also enables color output in fidl-lint
depending on the presence of the NO_COLOR env var and whether isatty
is true.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392793
Original-Revision: 1f6df840a4895611ca6e6e6a44cfa4d867fe23d7
GitOrigin-RevId: b8a0783c9ed5b6894ac99b6c9cb2d2f6830d9d08
Change-Id: Ieebebf0dd5517fc811821e7a610d997eb77967d9