[roll] Roll fuchsia [camera] Fix threading race in camera-gym BufferCollage.
This change fixes a race in the BufferCollage.AddCollection method.
Previously, the method made fidl API calls on the caller's thread, even
though the bindings were associated with the class's thread. This is
not supported by the hlcpp fidl bindings and can lead to crashes. This
change puts the bulk of AddCollection into a single task, and removes
the use of all but one fit::promise, opting for the more conventional
nested callback approach. This has a side-effect of serializing the
calls to Scenic and Sysmem, but this should not impact performance.

Test: Ran camera-gym and verified no behavior change.
      Ran camera-gym with cherry-picked fxr/385453.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392261
Original-Revision: 411fe50b9ecd321ab19438d50a9155a34f4eb2cb
GitOrigin-RevId: 4242dbea96021e61f3c887be7cadd77e52814509
Change-Id: Iae0310dc3e69b19fd36a23d04db5db90549aee60