[roll] Roll fuchsia [kernel][vm] Remove racy asserts
These asserts that confirm that a vm_page is in a list are well
intentioned, but ultimately racy. When these asserts are performed the
queue_node is owned by the page queue, and any usage of it needs to be
serialized with the page queue lock.

Although the page is guaranteed to be in a list, the page queue could
be moving the page between lists, whilst holding its lock. This makes
performing this assert, without holding the page queue lock, ultimately
racy.

Test: Exploiting this is difficult to do reliably as it requires
precise timing between multiple threads, so no tests added.

Original-Bug: 52716
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392496
Original-Revision: af5439aa33871bbfeb8dd54cb2f3c87b1c9c7354
GitOrigin-RevId: cd40bfb091bcc8e6b9ddb3742a87bef771a94782
Change-Id: Iaed1e5586a244a7607e33a2b66df417d693e31e4