[roll] Roll fuchsia [paver] Add mechanism to use bufferred client
This CL impelments a mechanism to swtich to the new SyncClientBuffered.
The change consists of the following:

1. A paver context object is introduced in the paver service to host
and share information and unique instance of objects among created
partition-clients.

2. A unique instance of SyncClientBufferred is created and hosted in the
paver context object. The context is passed to a new partition client--
AstroSysconfigPartitionClientBuffered for performing IO.

3. AstroPartitioner queries boot argument astro.sysconfig.io-optimization
to determine whether or not to use
AstroSysconfigPartitionClientBuffered.

Original-Bug: 47505
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382502
Original-Revision: cf53439d8b43395d997ac2bdbced357109f22288
GitOrigin-RevId: 653bb818fb554657e1be839f304acca79c6ec213
Change-Id: I3e50a6387c8cf8bba33b63bc1c139c7604df983e