[roll] Roll fuchsia [bt][sm] Minor patches to already-landed phase classes
V5.0 Vol. 3 Part H 3.6.1 says to ignore the EncKey bit for Secure
Connections pairing. This could be done by explicitly checking for
Secure Connections and ignoring the EncKey bit in the KeyDistGen fields
whenever they're accessed, but it is simpler to ignore the EncKey bit
by setting it to 0 when resolving the negotiated features in Phase 1.

Also adds an assertion to Phase 3 that it has not failed when its public
methods are called.

Original-Bug: 46278
Test: `fx test bt-host-unittests -- --gtest_filter="SMP*"`
      Added test SMP_Phase1Test.FeatureExchangeScIgnoresEncKeyBit.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386215
Original-Revision: d209097450bec007be47adc97b8b9dfd0107feea
GitOrigin-RevId: 6fdc490a6f05c1460c276c496fcb5a6bee351640
Change-Id: I0f18419073ae1367d6f4cd7f95c7a987de53b373