[roll] Roll fuchsia [debugger] Clarify symbol status.
Updates the symbol index when sym-stat is called. This ensures the index
stats are up-to-date with any modifications to the symbol path.
Previously, updates to the symbol-path was not reflected in the status
until something was loaded, which made diagnosing problems more
difficult.

Add text to the output of sym-stat indicating that the index was just
updated so there's no ambiguity.

Add error output to the symbol load status in all paths. Previously
there were some paths in loading files in the build ID index that didn't
write status.

Add more troubleshooting to the help.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375918
Original-Revision: e7b482c364a9219c55143dd67af8259793b828f0
GitOrigin-RevId: 2198ded71d0da820906389c12a954a1c823daf5b
Change-Id: Ie3d002ee2483ae78d15a257e4c3d1f0bdf296f6a