[roll] Roll fuchsia [debugger] Don't escape C++ destructor names.
Destructors were erroniously escaped for display, as in
"MyClass::$(~MyClass)" This is because the formatting code that checks
for escaping was different than the tokenizing code that knows about how
to parse names.

This change adds a new helper function to the tokenizer for the name
formatter to use to know whether something is a valid name and therefore
doesn't need escaping.

"$" was removed as a valid name character. Previously this was in there
but the case never got hit because "$" is handled as a special-case
before this code was called. Without this change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392416
Original-Revision: 3066194765680c42aa4147458553a13f1584a31c
GitOrigin-RevId: 7ae965a80bb7890953c64e49e6053f7a244a19ac
Change-Id: Icffb84ca4bd333d60c39d169e8811c23051a7a37