[roll] Roll fuchsia [SL4F] allow graceful exit of a component
The code was treating any component exit as a failure. This change accepts a
result code of 0 (widely accepted as a graceful shutdown/exit) as a non-error.
A non-zero exit code will continue to be reported as an unexpected exit.

Original-Bug: 52840
Test: fx run-e2e-tests sl4f_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391754
Original-Revision: 13d59376bfb5763363cecdd4ca7c123d70d8504f
GitOrigin-RevId: 93ff69931405a181ea74054264bbca43826210c3
Change-Id: I29c80e4993b7bed7b35226cde2f129cd1cd35e01