[roll] Roll fuchsia [cml] Introduce static EventStreams
We will be using the CapabilityRouted v2 event in order to do log
attribution of v2 components. However, a component can connect to
a LogSink before Archivist starts and so we must ensure that Archivist
starts in response to a CapabilityRouted event.

The way to do this is to enable starting a component in response to an
Event. Static EventStreams enable this.

This CL also deletes the golden tests which are exact
whitespace-for-whitespace tests which doesn't mesh well with formatting
of cm files.

Sample syntax:

{
    event_stream: [
        "began",
        "capability_ready",
        "destroyed",
    ],
    as: "/svc/my_stream",
},

Original-Bug: 51179
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390614
Original-Revision: b281132bdd517c41a4613b1ab3a2281372b236b3
GitOrigin-RevId: 7024f3bd2f3cba8da249349eb0c3b1948f8ff18a
Change-Id: I2fcd2d99f975787f684fe1f583b917152b9121e8