[roll] Roll fuchsia [cobalt][test_app] Removes a questionable test.
I am removing TestLogElapsedTimeWithAggregationWorkerRunning
because it seems like it is not the right way to test what
it was trying to test, and also it involved a sleep thereby
making the test potentially flaky and take a longer time.

For a reason I don't understand the test was repeated 10 times and
each run of the test involved a 1 second sleep. Based on the
error message in the failure case it seems the goal of the
test was: to check that after 1 second, the EventAggregatorManager
should only have run 1 loop. I think there are better ways to
test this--directly in Cobalt core. This integration test
involving FIDL doesn't seem like the right way to check this.

Also, I am having some trouble understanding the original
purpose of this test. Looking through code history the first
time I see this test is in fxr/358412 which is a re-apply of
a revert. So I can't understand the goal of this test. But
whatever the goal is, I bet that goal can be better met
with a test in Cobalt Core.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392538
Original-Revision: c42da406c787c7da79864e644513d13e4554ee4e
GitOrigin-RevId: 6f5cb97e297e0e664bccc6b1b3f2d5cc56711427
Change-Id: Ie3bdff30b42c659832e23bb2f3fcacabd8d29a5f