[roll] Roll fuchsia [zircon][tests] Do not assert on thread time not changing
The signal that a thread is terminated is returned before the thread's
final reschedule and accounting. It is not correct to assert that a dead
thread's cpu time does not change, since we do not make strong
guarantees about this ordering.

Original-Fixed: 52870
Original-Bug: 52872

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393090
Original-Revision: 5a66ce661c6e596f78cea9d0bdfe068339666be7
GitOrigin-RevId: 5f340c1d9be80652386711a77f239359139e78ce
Change-Id: I803672e440c02551a9a29f253cdad309b368761f