[roll] Roll fuchsia [ffx] Refactor select command for better FE UI.
This cleans up the structure we send between the FE and the daemon to
break the path down into consituent parts, so that they can be grouped
accordingly on the frontend.

This also moves `ffx select` into a plugin on the front-end side.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390814
Original-Revision: 67d9c7a85ed5aa07a697467d34dd765fa4d8f86f
GitOrigin-RevId: fd7a6e96d2ea6c5b5e15129a2aeb8b769df3c7c2
Change-Id: I15327d2a4aecb19fb279569d8f360c15dfc61b8c