[roll] Roll fuchsia [power_manager] Add DriverManagerHandler node
Implement a new node called DriverManagerHandler that is responsible for
serving the fuchsia.power.manager.DriverManagerRegistration service. The
Driver Manager will be the sole client of this protocol and will use it
to provide the Power Manager with the protocol instances that it
requires in order to drive orderly system shutdown. The new node will
not be instantiated in the Power Manager until it is specified in a node
config file.

Original-Bug: 51043
Original-Bug: 51078
Original-Bug: 51092
Test: fx test power_manager_tests
Multiply: power_manager_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391681
Original-Revision: b34a0679138a8aa7e94a2b38ea79e2960f636327
GitOrigin-RevId: 29f6652aff710c6dcdf563ec86222a668a4bf4c8
Change-Id: Iae7e3cb0a8148df678f0521a76f789769c1dc879