[roll] Roll fuchsia [blobstats] Made blobstats a proper host tool.
Initial refactoring of the blobstats code, starting with
building it as a host tool. That involves passing all Dart
static analysis, which was ignored before when blobstats
was run as a standalone script.
Also added the first couple of tests, now that the code
allows to add tests. At the same time, removed the empty
blobstats_host_test which is now replaced by the end-to-end
fx test added in the last CL.
Tested by running `fx blobstats` in addition to the unit tests
and the end-to-end test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391073
Original-Revision: 186f45a49c51276c496af16718de30c8da0ea54c
GitOrigin-RevId: 5269975cc84ee5fac2917f33dda59c8ae7de2213
Change-Id: I0bd7ba4c1a64b5a5de5f3973337f1999ce557a97