[roll] Roll fuchsia [FDF] Fix flake in rebind
Devices can be added as invisible/or a init hook can be implemented if a
driver has non trivial amount of initialization to do.
Device controller rebind api when called on parent, tears down the stack
and rebinds the driver. When device controller rebind is
called on the DevfsConnection the fidl txn is cached in zx_device_t and
is completed when the init task/thread completes. At the moment, when a
init task/thread completes we check for any cached rebind/bind fidl txns
and complete them.

The current race in the below
mentioned bug is like so:
1. Initially the driver's bind routine gets called automatically as
there is a driver that matches the parent device. This is before calling
the Rebind device controller api from the test. This is automatic
binding that happens by default.
2. The device added by the bind routine in step#1 is INVISIBLE, meaning
there is an init task/thread.
3. Before the init task/thread gets scheduled, the test issues rebind
on the devfs connection of the parent.
3. The rebind fidl txn is saved in the device to be replied later.
4. The init task/thread that was part of autobind from step 1
completes and the rebind device controller request issued by the test gets
wrongly completed.
This change solves the above race by making sure, we reply to
bind/rebind connections only if they exist when the bind of the device
gets called.In step1 the when bind routine gets called, as part of
autobind we will not save any cached txn in the device. We will use this
information to solve the race.

Original-Bug:49618

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392315
Original-Revision: 2111cae347121dab5975ac589ea74655983c142a
GitOrigin-RevId: 8545ff5c2de71c4a3b02678ba15f1fe725b5cd18
Change-Id: I7f9c3927240cd3a6aef838085b213d3fe5401442