[roll] Roll fuchsia [kernel][vm] Expand FindInitialPageContentLocked usage
This changes FindInitialPageContentLocked to be more general such that
it can be called even when there is already a committed page in the
current VMO, and it returns a reference to a VmPageOrMarker instead of
a vm_page.

Allowing to be called when there is already a committed page allows for
ZeroRangeLocked to use it determine what to do with a committed page.
This behavior was already being relied upon by ZeroRangeLocked, and
would trigger the assertion. The included unittest hits this path.

Returning a VmPageOrMarker allows for the caller to distinguish
between an explicit zero page that has been actively written, and an
absence of content that will typically (but not always) mean the
initial content is zero. The one scenario where the initial content
would be missing and not zero is when there is a page source. No
test is included for this as there is no legal way at the moment to
construct a VMO that has a page source and a marker, so the present
error in distinguishing cannot be observed.

The code in GetPageLocked is refactored to better take advantage of
the new return type.

Original-Bug: 52753
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392616
Original-Revision: dcc66882c6ad3747fcabbfdda222536c4129ba5c
GitOrigin-RevId: 37cd610e7acb578f029cbc3456157c509d681dc9
Change-Id: Ia2cca2f3b83a9bf8bf60fb015ee22ceef0bae6a1