[roll] Roll fuchsia [brcmfmac] Add MSGBUF ioctl support
This commit adds support for MSGBUF commands in the ioctl format.  This
is a command format derived from the Linux history of the driver; the
name "ioctl" is itself fairly meaningless on Fuchsia.

Functionally, this provides the implementations of set_dcmd() and
get_dcmd() in the brcmf_proto interface.

This commit also fixes a race in DmaPool::Buffer between buffer
releasing and re-acquiring, in the common use case that a thread passes
a buffer off to another thread before releasing it.

Original-Bug: 40719
Test: Booted and fx run-test brcmfmac_msgbuf_test on NUC
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/340889
Original-Revision: 90c8080c3e55c8aee1c297f755ae3a6091c16a6e
GitOrigin-RevId: c07aade4dbadc3cdad3df4c396f92842e36b9337
Change-Id: I25c4d8608cc7634af31756fdf218d18c3c92f559