[roll] Roll fuchsia [archivist] Fix dropped logs due to contended lock
If the component stats log is contended, the poll state of the lock was
returned as the poll state of the log stream. This would then cause the
log stream to be terminated early, and drop logs.

This state can occur when there is a large number of requests at once,
and all requests share the same source identity. This was true of all
driver log sink requests, that occur before archivist is started.

This change moves the stats handling outside of
LogMessageSocket::poll_next.

Original-Bug: 49457
Original-Bug: 52619
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392973
Original-Revision: a10985b8404a149db36feff20a5d3e829772a45d
GitOrigin-RevId: d17ab06e5a0175d26cfe65c3815775b408fd72a7
Change-Id: I0cac9c2baf86049866faf831e4a529039197e6ad