[roll] Roll fuchsia [paver] Add mechanism to use bufferred client
This CL impelments a mechanism to swtich to the new SyncClientBuffered.
The change consists of the following:

1. A paver context object is introduced in the paver service to host
and share information and unique instance of objects among created
partition-clients.

2. A unique instance of SyncClientBufferred is created and hosted in the
paver context object. The context is passed to a new partition client--
AstroSysconfigPartitionClientBuffered for performing IO.

3. AstroPartitioner queries boot argument to determine whether or not to
use AstroSysconfigPartitionClientBuffered.

The CL is an update on reverted CL http://fxr/382502. The main changes
are:

1. A temporary new boot arg astro.sysconfig.buffered-client is
introduced to control whether buffered client is enabled at all. It is
disabled by default.
2. Unittest are updated accordingly.

Original-Bug: 47505
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393334
Original-Revision: 77dd15732cfe04d69ba7594f6dfa4afa2af84042
GitOrigin-RevId: 039e9b54de825dd83f4adfd10158e3ec58c163dd
Change-Id: Iacfc62ef5501a05337c4b01a6317f60980b52c20