[roll] Roll fuchsia [sysmem] Fix AFBC format modifier capitalization.
They need to be all-caps, so the lower-case "x" was incorrect. They're
not used outside src/, so we don't need a soft transition.

There are similar Magma format modifier values that we also need to
fix the capitalization of. They're used in third-party code, so we need
to soft-transition them.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393086
Original-Revision: ca46e8caac6d3581f49fd9119a737d56eed98d6c
GitOrigin-RevId: 3cc8f54ad83c6543f0505575f53842b9726f3990
Change-Id: I7378164cf9d5422e263e6cedd18bfb3a8fd255a6